Struct aws_sdk_elasticache::client::fluent_builders::CreateCacheCluster

pub struct CreateCacheCluster<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateCacheCluster<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateCacheClusterOutput, SdkError<CreateCacheClusterError>> where
R::Policy: SmithyRetryPolicy<CreateCacheClusterInputOperationOutputAlias, CreateCacheClusterOutput, CreateCacheClusterError, CreateCacheClusterInputOperationRetryAlias>,

pub fn cache_cluster_id(self, inp: impl Into<String>) -> Self

The node group (shard) identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 50 alphanumeric characters or hyphens.

The first character must be a letter.

A name cannot end with a hyphen or contain two consecutive hyphens.

pub fn set_cache_cluster_id(self, input: Option<String>) -> Self

pub fn replication_group_id(self, inp: impl Into<String>) -> Self

The ID of the replication group to which this cluster should belong. If this parameter is specified, the cluster is added to the specified replication group as a read replica; otherwise, the cluster is a standalone primary that is not part of any replication group.

If the specified replication group is Multi-AZ enabled and the Availability Zone is not specified, the cluster is created in Availability Zones that provide the best spread of read replicas across Availability Zones.

This parameter is only valid if the Engine parameter is redis.

pub fn set_replication_group_id(self, input: Option<String>) -> Self

pub fn az_mode(self, inp: AzMode) -> Self

Specifies whether the nodes in this Memcached cluster are created in a single Availability Zone or created across multiple Availability Zones in the cluster's region.

This parameter is only supported for Memcached clusters.

If the AZMode and PreferredAvailabilityZones are not specified, ElastiCache assumes single-az mode.

pub fn set_az_mode(self, input: Option<AzMode>) -> Self

pub fn preferred_availability_zone(self, inp: impl Into<String>) -> Self

The EC2 Availability Zone in which the cluster is created.

All nodes belonging to this cluster are placed in the preferred Availability Zone. If you want to create your nodes across multiple Availability Zones, use PreferredAvailabilityZones.

Default: System chosen Availability Zone.

pub fn set_preferred_availability_zone(self, input: Option<String>) -> Self

pub fn preferred_availability_zones(self, inp: impl Into<String>) -> Self

Appends an item to PreferredAvailabilityZones.

To override the contents of this collection use set_preferred_availability_zones.

A list of the Availability Zones in which cache nodes are created. The order of the zones in the list is not important.

This option is only supported on Memcached.

If you are creating your cluster in an Amazon VPC (recommended) you can only locate nodes in Availability Zones that are associated with the subnets in the selected subnet group.

The number of Availability Zones listed must equal the value of NumCacheNodes.

If you want all the nodes in the same Availability Zone, use PreferredAvailabilityZone instead, or repeat the Availability Zone multiple times in the list.

Default: System chosen Availability Zones.

pub fn set_preferred_availability_zones(
self,
input: Option<Vec<String>>
) -> Self

pub fn num_cache_nodes(self, inp: i32) -> Self

The initial number of cache nodes that the cluster has.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 40.

If you need more than 40 nodes for your Memcached cluster, please fill out the ElastiCache Limit Increase Request form at http://aws.amazon.com/contact-us/elasticache-node-limit-request/.

pub fn set_num_cache_nodes(self, input: Option<i32>) -> Self

pub fn cache_node_type(self, inp: impl Into<String>) -> Self

The compute and memory capacity of the nodes in the node group (shard).

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

pub fn set_cache_node_type(self, input: Option<String>) -> Self

pub fn engine(self, inp: impl Into<String>) -> Self

The name of the cache engine to be used for this cluster.

Valid values for this parameter are: memcached | redis

pub fn set_engine(self, input: Option<String>) -> Self

pub fn engine_version(self, inp: impl Into<String>) -> Self

The version number of the cache engine to be used for this cluster. To view the supported cache engine versions, use the DescribeCacheEngineVersions operation.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cluster or replication group and create it anew with the earlier engine version.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn cache_parameter_group_name(self, inp: impl Into<String>) -> Self

The name of the parameter group to associate with this cluster. If this argument is omitted, the default parameter group for the specified engine is used. You cannot use any parameter group which has cluster-enabled='yes' when creating a cluster.

pub fn set_cache_parameter_group_name(self, input: Option<String>) -> Self

pub fn cache_subnet_group_name(self, inp: impl Into<String>) -> Self

The name of the subnet group to be used for the cluster.

Use this parameter only when you are creating a cluster in an Amazon Virtual Private Cloud (Amazon VPC).

If you're going to launch your cluster in an Amazon VPC, you need to create a subnet group before you start creating a cluster. For more information, see Subnets and Subnet Groups.

pub fn set_cache_subnet_group_name(self, input: Option<String>) -> Self

pub fn cache_security_group_names(self, inp: impl Into<String>) -> Self

Appends an item to CacheSecurityGroupNames.

To override the contents of this collection use set_cache_security_group_names.

A list of security group names to associate with this cluster.

Use this parameter only when you are creating a cluster outside of an Amazon Virtual Private Cloud (Amazon VPC).

pub fn set_cache_security_group_names(self, input: Option<Vec<String>>) -> Self

pub fn security_group_ids(self, inp: impl Into<String>) -> Self

Appends an item to SecurityGroupIds.

To override the contents of this collection use set_security_group_ids.

One or more VPC security groups associated with the cluster.

Use this parameter only when you are creating a cluster in an Amazon Virtual Private Cloud (Amazon VPC).

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

A list of tags to be added to this resource.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn snapshot_arns(self, inp: impl Into<String>) -> Self

Appends an item to SnapshotArns.

To override the contents of this collection use set_snapshot_arns.

A single-element string list containing an Amazon Resource Name (ARN) that uniquely identifies a Redis RDB snapshot file stored in Amazon S3. The snapshot file is used to populate the node group (shard). The Amazon S3 object name in the ARN cannot contain any commas.

This parameter is only valid if the Engine parameter is redis.

Example of an Amazon S3 ARN: arn:aws:s3:::my_bucket/snapshot1.rdb

pub fn set_snapshot_arns(self, input: Option<Vec<String>>) -> Self

pub fn snapshot_name(self, inp: impl Into<String>) -> Self

The name of a Redis snapshot from which to restore data into the new node group (shard). The snapshot status changes to restoring while the new node group (shard) is being created.

This parameter is only valid if the Engine parameter is redis.

pub fn set_snapshot_name(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, inp: impl Into<String>) -> Self

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn port(self, inp: i32) -> Self

The port number on which each of the cache nodes accepts connections.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn notification_topic_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

The Amazon SNS topic owner must be the same as the cluster owner.

pub fn set_notification_topic_arn(self, input: Option<String>) -> Self

pub fn auto_minor_version_upgrade(self, inp: bool) -> Self

This parameter is currently disabled.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn snapshot_retention_limit(self, inp: i32) -> Self

The number of days for which ElastiCache retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot taken today is retained for 5 days before being deleted.

This parameter is only valid if the Engine parameter is redis.

Default: 0 (i.e., automatic backups are disabled for this cache cluster).

pub fn set_snapshot_retention_limit(self, input: Option<i32>) -> Self

pub fn snapshot_window(self, inp: impl Into<String>) -> Self

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard).

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

This parameter is only valid if the Engine parameter is redis.

pub fn set_snapshot_window(self, input: Option<String>) -> Self

pub fn auth_token(self, inp: impl Into<String>) -> Self

Reserved parameter. The password used to access a password protected server.

Password constraints:

Must be only printable ASCII characters.

Must be at least 16 characters and no more than 128 characters in length.

The only permitted printable special characters are !, &, #, $, ^, <, >, and -. Other printable special characters cannot be used in the AUTH token.

For more information, see AUTH password at http://redis.io/commands/AUTH.

pub fn set_auth_token(self, input: Option<String>) -> Self

pub fn outpost_mode(self, inp: OutpostMode) -> Self

Specifies whether the nodes in the cluster are created in a single outpost or across multiple outposts.

pub fn set_outpost_mode(self, input: Option<OutpostMode>) -> Self

pub fn preferred_outpost_arn(self, inp: impl Into<String>) -> Self

The outpost ARN in which the cache cluster is created.

pub fn set_preferred_outpost_arn(self, input: Option<String>) -> Self

pub fn preferred_outpost_arns(self, inp: impl Into<String>) -> Self

Appends an item to PreferredOutpostArns.

To override the contents of this collection use set_preferred_outpost_arns.

The outpost ARNs in which the cache cluster is created.

pub fn set_preferred_outpost_arns(self, input: Option<Vec<String>>) -> Self

pub fn log_delivery_configurations(
self,
inp: impl Into<LogDeliveryConfigurationRequest>
) -> Self

Appends an item to LogDeliveryConfigurations.

To override the contents of this collection use set_log_delivery_configurations.

Specifies the destination, format and type of the logs.

pub fn set_log_delivery_configurations(
self,
input: Option<Vec<LogDeliveryConfigurationRequest>>
) -> Self