Struct aws_sdk_elasticache::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonElastiCacheV9.

This client allows ergonomic access to a AmazonElastiCacheV9-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_tags_to_resource(&self) -> AddTagsToResource<C, M, R>

pub fn authorize_cache_security_group_ingress(
&self
) -> AuthorizeCacheSecurityGroupIngress<C, M, R>

pub fn batch_apply_update_action(&self) -> BatchApplyUpdateAction<C, M, R>

pub fn batch_stop_update_action(&self) -> BatchStopUpdateAction<C, M, R>

pub fn complete_migration(&self) -> CompleteMigration<C, M, R>

pub fn copy_snapshot(&self) -> CopySnapshot<C, M, R>

pub fn create_cache_cluster(&self) -> CreateCacheCluster<C, M, R>

pub fn create_cache_parameter_group(&self) -> CreateCacheParameterGroup<C, M, R>

pub fn create_cache_security_group(&self) -> CreateCacheSecurityGroup<C, M, R>

pub fn create_cache_subnet_group(&self) -> CreateCacheSubnetGroup<C, M, R>

pub fn create_global_replication_group(
&self
) -> CreateGlobalReplicationGroup<C, M, R>

pub fn create_replication_group(&self) -> CreateReplicationGroup<C, M, R>

pub fn create_snapshot(&self) -> CreateSnapshot<C, M, R>

pub fn create_user(&self) -> CreateUser<C, M, R>

pub fn create_user_group(&self) -> CreateUserGroup<C, M, R>

pub fn decrease_node_groups_in_global_replication_group(
&self
) -> DecreaseNodeGroupsInGlobalReplicationGroup<C, M, R>

pub fn decrease_replica_count(&self) -> DecreaseReplicaCount<C, M, R>

pub fn delete_cache_cluster(&self) -> DeleteCacheCluster<C, M, R>

pub fn delete_cache_parameter_group(&self) -> DeleteCacheParameterGroup<C, M, R>

pub fn delete_cache_security_group(&self) -> DeleteCacheSecurityGroup<C, M, R>

pub fn delete_cache_subnet_group(&self) -> DeleteCacheSubnetGroup<C, M, R>

pub fn delete_global_replication_group(
&self
) -> DeleteGlobalReplicationGroup<C, M, R>

pub fn delete_replication_group(&self) -> DeleteReplicationGroup<C, M, R>

pub fn delete_snapshot(&self) -> DeleteSnapshot<C, M, R>

pub fn delete_user(&self) -> DeleteUser<C, M, R>

pub fn delete_user_group(&self) -> DeleteUserGroup<C, M, R>

pub fn describe_cache_clusters(&self) -> DescribeCacheClusters<C, M, R>

pub fn describe_cache_engine_versions(
&self
) -> DescribeCacheEngineVersions<C, M, R>

pub fn describe_cache_parameter_groups(
&self
) -> DescribeCacheParameterGroups<C, M, R>

pub fn describe_cache_parameters(&self) -> DescribeCacheParameters<C, M, R>

pub fn describe_cache_security_groups(
&self
) -> DescribeCacheSecurityGroups<C, M, R>

pub fn describe_cache_subnet_groups(&self) -> DescribeCacheSubnetGroups<C, M, R>

pub fn describe_engine_default_parameters(
&self
) -> DescribeEngineDefaultParameters<C, M, R>

pub fn describe_events(&self) -> DescribeEvents<C, M, R>

pub fn describe_global_replication_groups(
&self
) -> DescribeGlobalReplicationGroups<C, M, R>

pub fn describe_replication_groups(&self) -> DescribeReplicationGroups<C, M, R>

pub fn describe_reserved_cache_nodes(
&self
) -> DescribeReservedCacheNodes<C, M, R>

pub fn describe_reserved_cache_nodes_offerings(
&self
) -> DescribeReservedCacheNodesOfferings<C, M, R>

pub fn describe_service_updates(&self) -> DescribeServiceUpdates<C, M, R>

pub fn describe_snapshots(&self) -> DescribeSnapshots<C, M, R>

pub fn describe_update_actions(&self) -> DescribeUpdateActions<C, M, R>

pub fn describe_user_groups(&self) -> DescribeUserGroups<C, M, R>

pub fn describe_users(&self) -> DescribeUsers<C, M, R>

pub fn disassociate_global_replication_group(
&self
) -> DisassociateGlobalReplicationGroup<C, M, R>

pub fn failover_global_replication_group(
&self
) -> FailoverGlobalReplicationGroup<C, M, R>

pub fn increase_node_groups_in_global_replication_group(
&self
) -> IncreaseNodeGroupsInGlobalReplicationGroup<C, M, R>

pub fn increase_replica_count(&self) -> IncreaseReplicaCount<C, M, R>

pub fn list_allowed_node_type_modifications(
&self
) -> ListAllowedNodeTypeModifications<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn modify_cache_cluster(&self) -> ModifyCacheCluster<C, M, R>

pub fn modify_cache_parameter_group(&self) -> ModifyCacheParameterGroup<C, M, R>

pub fn modify_cache_subnet_group(&self) -> ModifyCacheSubnetGroup<C, M, R>

pub fn modify_global_replication_group(
&self
) -> ModifyGlobalReplicationGroup<C, M, R>

pub fn modify_replication_group(&self) -> ModifyReplicationGroup<C, M, R>

pub fn modify_replication_group_shard_configuration(
&self
) -> ModifyReplicationGroupShardConfiguration<C, M, R>

pub fn modify_user(&self) -> ModifyUser<C, M, R>

pub fn modify_user_group(&self) -> ModifyUserGroup<C, M, R>

pub fn purchase_reserved_cache_nodes_offering(
&self
) -> PurchaseReservedCacheNodesOffering<C, M, R>

pub fn rebalance_slots_in_global_replication_group(
&self
) -> RebalanceSlotsInGlobalReplicationGroup<C, M, R>

pub fn reboot_cache_cluster(&self) -> RebootCacheCluster<C, M, R>

pub fn remove_tags_from_resource(&self) -> RemoveTagsFromResource<C, M, R>

pub fn reset_cache_parameter_group(&self) -> ResetCacheParameterGroup<C, M, R>

pub fn revoke_cache_security_group_ingress(
&self
) -> RevokeCacheSecurityGroupIngress<C, M, R>

pub fn start_migration(&self) -> StartMigration<C, M, R>

pub fn test_failover(&self) -> TestFailover<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self