Module aws_sdk_elasticache::error

Modules

api_call_rate_for_customer_exceeded_fault

See ApiCallRateForCustomerExceededFault

authorization_already_exists_fault

See AuthorizationAlreadyExistsFault

authorization_not_found_fault

See AuthorizationNotFoundFault

cache_cluster_already_exists_fault

See CacheClusterAlreadyExistsFault

cache_cluster_not_found_fault

See CacheClusterNotFoundFault

cache_parameter_group_already_exists_fault

See CacheParameterGroupAlreadyExistsFault

cache_parameter_group_not_found_fault

See CacheParameterGroupNotFoundFault

cache_parameter_group_quota_exceeded_fault

See CacheParameterGroupQuotaExceededFault

cache_security_group_already_exists_fault

See CacheSecurityGroupAlreadyExistsFault

cache_security_group_not_found_fault

See CacheSecurityGroupNotFoundFault

cache_security_group_quota_exceeded_fault

See CacheSecurityGroupQuotaExceededFault

cache_subnet_group_already_exists_fault

See CacheSubnetGroupAlreadyExistsFault

cache_subnet_group_in_use

See CacheSubnetGroupInUse

cache_subnet_group_not_found_fault

See CacheSubnetGroupNotFoundFault

cache_subnet_group_quota_exceeded_fault

See CacheSubnetGroupQuotaExceededFault

cache_subnet_quota_exceeded_fault

See CacheSubnetQuotaExceededFault

cluster_quota_for_customer_exceeded_fault

See ClusterQuotaForCustomerExceededFault

default_user_associated_to_user_group_fault

See DefaultUserAssociatedToUserGroupFault

default_user_required

See DefaultUserRequired

duplicate_user_name_fault

See DuplicateUserNameFault

global_replication_group_already_exists_fault

See GlobalReplicationGroupAlreadyExistsFault

global_replication_group_not_found_fault

See GlobalReplicationGroupNotFoundFault

insufficient_cache_cluster_capacity_fault

See InsufficientCacheClusterCapacityFault

invalid_arn_fault

See InvalidArnFault

invalid_cache_cluster_state_fault

See InvalidCacheClusterStateFault

invalid_cache_parameter_group_state_fault

See InvalidCacheParameterGroupStateFault

invalid_cache_security_group_state_fault

See InvalidCacheSecurityGroupStateFault

invalid_global_replication_group_state_fault

See InvalidGlobalReplicationGroupStateFault

invalid_kms_key_fault

See InvalidKmsKeyFault

invalid_parameter_combination_exception

See InvalidParameterCombinationException

invalid_parameter_value_exception

See InvalidParameterValueException

invalid_replication_group_state_fault

See InvalidReplicationGroupStateFault

invalid_snapshot_state_fault

See InvalidSnapshotStateFault

invalid_subnet

See InvalidSubnet

invalid_user_group_state_fault

See InvalidUserGroupStateFault

invalid_user_state_fault

See InvalidUserStateFault

invalid_vpc_network_state_fault

See InvalidVpcNetworkStateFault

no_operation_fault

See NoOperationFault

node_group_not_found_fault

See NodeGroupNotFoundFault

node_groups_per_replication_group_quota_exceeded_fault

See NodeGroupsPerReplicationGroupQuotaExceededFault

node_quota_for_cluster_exceeded_fault

See NodeQuotaForClusterExceededFault

node_quota_for_customer_exceeded_fault

See NodeQuotaForCustomerExceededFault

replication_group_already_exists_fault

See ReplicationGroupAlreadyExistsFault

replication_group_already_under_migration_fault

See ReplicationGroupAlreadyUnderMigrationFault

replication_group_not_found_fault

See ReplicationGroupNotFoundFault

replication_group_not_under_migration_fault

See ReplicationGroupNotUnderMigrationFault

reserved_cache_node_already_exists_fault

See ReservedCacheNodeAlreadyExistsFault

reserved_cache_node_not_found_fault

See ReservedCacheNodeNotFoundFault

reserved_cache_node_quota_exceeded_fault

See ReservedCacheNodeQuotaExceededFault

reserved_cache_nodes_offering_not_found_fault

See ReservedCacheNodesOfferingNotFoundFault

service_linked_role_not_found_fault

See ServiceLinkedRoleNotFoundFault

service_update_not_found_fault

See ServiceUpdateNotFoundFault

snapshot_already_exists_fault

See SnapshotAlreadyExistsFault

snapshot_feature_not_supported_fault

See SnapshotFeatureNotSupportedFault

snapshot_not_found_fault

See SnapshotNotFoundFault

snapshot_quota_exceeded_fault

See SnapshotQuotaExceededFault

subnet_in_use

See SubnetInUse

subnet_not_allowed_fault

See SubnetNotAllowedFault

tag_not_found_fault

See TagNotFoundFault

tag_quota_per_resource_exceeded

See TagQuotaPerResourceExceeded

test_failover_not_available_fault

See TestFailoverNotAvailableFault

user_already_exists_fault

See UserAlreadyExistsFault

user_group_already_exists_fault

See UserGroupAlreadyExistsFault

user_group_not_found_fault

See UserGroupNotFoundFault

user_group_quota_exceeded_fault

See UserGroupQuotaExceededFault

user_not_found_fault

See UserNotFoundFault

user_quota_exceeded_fault

See UserQuotaExceededFault

Structs

AddTagsToResourceError

ApiCallRateForCustomerExceededFault

The customer has exceeded the allowed rate of API calls.

AuthorizationAlreadyExistsFault

The specified Amazon EC2 security group is already authorized for the specified cache security group.

AuthorizationNotFoundFault

The specified Amazon EC2 security group is not authorized for the specified cache security group.

AuthorizeCacheSecurityGroupIngressError

BatchApplyUpdateActionError

BatchStopUpdateActionError

CacheClusterAlreadyExistsFault

You already have a cluster with the given identifier.

CacheClusterNotFoundFault

The requested cluster ID does not refer to an existing cluster.

CacheParameterGroupAlreadyExistsFault

A cache parameter group with the requested name already exists.

CacheParameterGroupNotFoundFault

The requested cache parameter group name does not refer to an existing cache parameter group.

CacheParameterGroupQuotaExceededFault

The request cannot be processed because it would exceed the maximum number of cache security groups.

CacheSecurityGroupAlreadyExistsFault

A cache security group with the specified name already exists.

CacheSecurityGroupNotFoundFault

The requested cache security group name does not refer to an existing cache security group.

CacheSecurityGroupQuotaExceededFault

The request cannot be processed because it would exceed the allowed number of cache security groups.

CacheSubnetGroupAlreadyExistsFault

The requested cache subnet group name is already in use by an existing cache subnet group.

CacheSubnetGroupInUse

The requested cache subnet group is currently in use.

CacheSubnetGroupNotFoundFault

The requested cache subnet group name does not refer to an existing cache subnet group.

CacheSubnetGroupQuotaExceededFault

The request cannot be processed because it would exceed the allowed number of cache subnet groups.

CacheSubnetQuotaExceededFault

The request cannot be processed because it would exceed the allowed number of subnets in a cache subnet group.

ClusterQuotaForCustomerExceededFault

The request cannot be processed because it would exceed the allowed number of clusters per customer.

CompleteMigrationError

CopySnapshotError

CreateCacheClusterError

CreateCacheParameterGroupError

CreateCacheSecurityGroupError

CreateCacheSubnetGroupError

CreateGlobalReplicationGroupError

CreateReplicationGroupError

CreateSnapshotError

CreateUserError

CreateUserGroupError

DecreaseNodeGroupsInGlobalReplicationGroupError

DecreaseReplicaCountError

DefaultUserAssociatedToUserGroupFault

The default user assigned to the user group.

DefaultUserRequired

You must add default user to a user group.

DeleteCacheClusterError

DeleteCacheParameterGroupError

DeleteCacheSecurityGroupError

DeleteCacheSubnetGroupError

DeleteGlobalReplicationGroupError

DeleteReplicationGroupError

DeleteSnapshotError

DeleteUserError

DeleteUserGroupError

DescribeCacheClustersError

DescribeCacheEngineVersionsError

DescribeCacheParameterGroupsError

DescribeCacheParametersError

DescribeCacheSecurityGroupsError

DescribeCacheSubnetGroupsError

DescribeEngineDefaultParametersError

DescribeEventsError

DescribeGlobalReplicationGroupsError

DescribeReplicationGroupsError

DescribeReservedCacheNodesError

DescribeReservedCacheNodesOfferingsError

DescribeServiceUpdatesError

DescribeSnapshotsError

DescribeUpdateActionsError

DescribeUserGroupsError

DescribeUsersError

DisassociateGlobalReplicationGroupError

DuplicateUserNameFault

A user with this username already exists.

FailoverGlobalReplicationGroupError

GlobalReplicationGroupAlreadyExistsFault

The Global datastore name already exists.

GlobalReplicationGroupNotFoundFault

The Global datastore does not exist

IncreaseNodeGroupsInGlobalReplicationGroupError

IncreaseReplicaCountError

InsufficientCacheClusterCapacityFault

The requested cache node type is not available in the specified Availability Zone.

InvalidArnFault

The requested Amazon Resource Name (ARN) does not refer to an existing resource.

InvalidCacheClusterStateFault

The requested cluster is not in the available state.

InvalidCacheParameterGroupStateFault

The current state of the cache parameter group does not allow the requested operation to occur.

InvalidCacheSecurityGroupStateFault

The current state of the cache security group does not allow deletion.

InvalidGlobalReplicationGroupStateFault

The Global datastore is not available or in primary-only state.

InvalidKmsKeyFault

The KMS key supplied is not valid.

InvalidParameterCombinationException

Two or more incompatible parameters were specified.

InvalidParameterValueException

The value for a parameter is invalid.

InvalidReplicationGroupStateFault

The requested replication group is not in the available state.

InvalidSnapshotStateFault

The current state of the snapshot does not allow the requested operation to occur.

InvalidSubnet

An invalid subnet identifier was specified.

InvalidUserGroupStateFault

The user group is not in an active state.

InvalidUserStateFault

The user is not in active state.

InvalidVpcNetworkStateFault

The VPC network is in an invalid state.

ListAllowedNodeTypeModificationsError

ListTagsForResourceError

ModifyCacheClusterError

ModifyCacheParameterGroupError

ModifyCacheSubnetGroupError

ModifyGlobalReplicationGroupError

ModifyReplicationGroupError

ModifyReplicationGroupShardConfigurationError

ModifyUserError

ModifyUserGroupError

NoOperationFault

The operation was not performed because no changes were required.

NodeGroupNotFoundFault

The node group specified by the NodeGroupId parameter could not be found.

NodeGroupsPerReplicationGroupQuotaExceededFault

The request cannot be processed because it would exceed the maximum allowed number

NodeQuotaForClusterExceededFault

The request cannot be processed because it would exceed the allowed number of cache nodes in a single cluster.

NodeQuotaForCustomerExceededFault

The request cannot be processed because it would exceed the allowed number of cache nodes per customer.

PurchaseReservedCacheNodesOfferingError

RebalanceSlotsInGlobalReplicationGroupError

RebootCacheClusterError

RemoveTagsFromResourceError

ReplicationGroupAlreadyExistsFault

The specified replication group already exists.

ReplicationGroupAlreadyUnderMigrationFault

The targeted replication group is not available.

ReplicationGroupNotFoundFault

The specified replication group does not exist.

ReplicationGroupNotUnderMigrationFault

The designated replication group is not available for data migration.

ReservedCacheNodeAlreadyExistsFault

You already have a reservation with the given identifier.

ReservedCacheNodeNotFoundFault

The requested reserved cache node was not found.

ReservedCacheNodeQuotaExceededFault

The request cannot be processed because it would exceed the user's cache node quota.

ReservedCacheNodesOfferingNotFoundFault

The requested cache node offering does not exist.

ResetCacheParameterGroupError

RevokeCacheSecurityGroupIngressError

ServiceLinkedRoleNotFoundFault

The specified service linked role (SLR) was not found.

ServiceUpdateNotFoundFault

The service update doesn't exist

SnapshotAlreadyExistsFault

You already have a snapshot with the given name.

SnapshotFeatureNotSupportedFault

You attempted one of the following operations:

SnapshotNotFoundFault

The requested snapshot name does not refer to an existing snapshot.

SnapshotQuotaExceededFault

The request cannot be processed because it would exceed the maximum number of snapshots.

StartMigrationError

SubnetInUse

The requested subnet is being used by another cache subnet group.

SubnetNotAllowedFault

At least one subnet ID does not match the other subnet IDs. This mismatch typically occurs when a

TagNotFoundFault

The requested tag was not found on this resource.

TagQuotaPerResourceExceeded

The request cannot be processed because it would cause the resource to have more than the allowed number of tags. The maximum number of tags permitted on a resource is 50.

TestFailoverError

TestFailoverNotAvailableFault

The TestFailover action is not available.

UserAlreadyExistsFault

A user with this ID already exists.

UserGroupAlreadyExistsFault

The user group with this ID already exists.

UserGroupNotFoundFault

The user group was not found or does not exist

UserGroupQuotaExceededFault

The number of users exceeds the user group limit.

UserNotFoundFault

The user does not exist or could not be found.

UserQuotaExceededFault

The quota of users has been exceeded.

Enums

AddTagsToResourceErrorKind

AuthorizeCacheSecurityGroupIngressErrorKind

BatchApplyUpdateActionErrorKind

BatchStopUpdateActionErrorKind

CompleteMigrationErrorKind

CopySnapshotErrorKind

CreateCacheClusterErrorKind

CreateCacheParameterGroupErrorKind

CreateCacheSecurityGroupErrorKind

CreateCacheSubnetGroupErrorKind

CreateGlobalReplicationGroupErrorKind

CreateReplicationGroupErrorKind

CreateSnapshotErrorKind

CreateUserErrorKind

CreateUserGroupErrorKind

DecreaseNodeGroupsInGlobalReplicationGroupErrorKind

DecreaseReplicaCountErrorKind

DeleteCacheClusterErrorKind

DeleteCacheParameterGroupErrorKind

DeleteCacheSecurityGroupErrorKind

DeleteCacheSubnetGroupErrorKind

DeleteGlobalReplicationGroupErrorKind

DeleteReplicationGroupErrorKind

DeleteSnapshotErrorKind

DeleteUserErrorKind

DeleteUserGroupErrorKind

DescribeCacheClustersErrorKind

DescribeCacheEngineVersionsErrorKind

DescribeCacheParameterGroupsErrorKind

DescribeCacheParametersErrorKind

DescribeCacheSecurityGroupsErrorKind

DescribeCacheSubnetGroupsErrorKind

DescribeEngineDefaultParametersErrorKind

DescribeEventsErrorKind

DescribeGlobalReplicationGroupsErrorKind

DescribeReplicationGroupsErrorKind

DescribeReservedCacheNodesErrorKind

DescribeReservedCacheNodesOfferingsErrorKind

DescribeServiceUpdatesErrorKind

DescribeSnapshotsErrorKind

DescribeUpdateActionsErrorKind

DescribeUserGroupsErrorKind

DescribeUsersErrorKind

DisassociateGlobalReplicationGroupErrorKind

FailoverGlobalReplicationGroupErrorKind

IncreaseNodeGroupsInGlobalReplicationGroupErrorKind

IncreaseReplicaCountErrorKind

ListAllowedNodeTypeModificationsErrorKind

ListTagsForResourceErrorKind

ModifyCacheClusterErrorKind

ModifyCacheParameterGroupErrorKind

ModifyCacheSubnetGroupErrorKind

ModifyGlobalReplicationGroupErrorKind

ModifyReplicationGroupErrorKind

ModifyReplicationGroupShardConfigurationErrorKind

ModifyUserErrorKind

ModifyUserGroupErrorKind

PurchaseReservedCacheNodesOfferingErrorKind

RebalanceSlotsInGlobalReplicationGroupErrorKind

RebootCacheClusterErrorKind

RemoveTagsFromResourceErrorKind

ResetCacheParameterGroupErrorKind

RevokeCacheSecurityGroupIngressErrorKind

StartMigrationErrorKind

TestFailoverErrorKind