Struct aws_sdk_elasticache::input::create_replication_group_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateReplicationGroupInput

Implementations

impl Builder

pub fn replication_group_id(self, input: impl Into<String>) -> Self

The replication group identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 40 alphanumeric characters or hyphens.

The first character must be a letter.

A name cannot end with a hyphen or contain two consecutive hyphens.

pub fn set_replication_group_id(self, input: Option<String>) -> Self

pub fn replication_group_description(self, input: impl Into<String>) -> Self

A user-created description for the replication group.

pub fn set_replication_group_description(self, input: Option<String>) -> Self

pub fn global_replication_group_id(self, input: impl Into<String>) -> Self

The name of the Global datastore

pub fn set_global_replication_group_id(self, input: Option<String>) -> Self

pub fn primary_cluster_id(self, input: impl Into<String>) -> Self

The identifier of the cluster that serves as the primary for this replication group. This cluster must already exist and have a status of available.

This parameter is not required if NumCacheClusters, NumNodeGroups, or ReplicasPerNodeGroup is specified.

pub fn set_primary_cluster_id(self, input: Option<String>) -> Self

pub fn automatic_failover_enabled(self, input: bool) -> Self

Specifies whether a read-only replica is automatically promoted to read/write primary if the existing primary fails.

AutomaticFailoverEnabled must be enabled for Redis (cluster mode enabled) replication groups.

Default: false

pub fn set_automatic_failover_enabled(self, input: Option<bool>) -> Self

pub fn multi_az_enabled(self, input: bool) -> Self

A flag indicating if you have Multi-AZ enabled to enhance fault tolerance. For more information, see Minimizing Downtime: Multi-AZ.

pub fn set_multi_az_enabled(self, input: Option<bool>) -> Self

pub fn num_cache_clusters(self, input: i32) -> Self

The number of clusters this replication group initially has.

This parameter is not used if there is more than one node group (shard). You should use ReplicasPerNodeGroup instead.

If AutomaticFailoverEnabled is true, the value of this parameter must be at least 2. If AutomaticFailoverEnabled is false you can omit this parameter (it will default to 1), or you can explicitly set it to a value between 2 and 6.

The maximum permitted value for NumCacheClusters is 6 (1 primary plus 5 replicas).

pub fn set_num_cache_clusters(self, input: Option<i32>) -> Self

pub fn preferred_cache_cluster_a_zs(self, input: impl Into<String>) -> Self

pub fn set_preferred_cache_cluster_a_zs(
self,
input: Option<Vec<String>>
) -> Self

pub fn num_node_groups(self, input: i32) -> Self

An optional parameter that specifies the number of node groups (shards) for this Redis (cluster mode enabled) replication group. For Redis (cluster mode disabled) either omit this parameter or set it to 1.

Default: 1

pub fn set_num_node_groups(self, input: Option<i32>) -> Self

pub fn replicas_per_node_group(self, input: i32) -> Self

An optional parameter that specifies the number of replica nodes in each node group (shard). Valid values are 0 to 5.

pub fn set_replicas_per_node_group(self, input: Option<i32>) -> Self

pub fn node_group_configuration(
self,
input: impl Into<NodeGroupConfiguration>
) -> Self

pub fn set_node_group_configuration(
self,
input: Option<Vec<NodeGroupConfiguration>>
) -> Self

pub fn cache_node_type(self, input: impl Into<String>) -> Self

The compute and memory capacity of the nodes in the node group (shard).

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

pub fn set_cache_node_type(self, input: Option<String>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

The name of the cache engine to be used for the clusters in this replication group. Must be Redis.

pub fn set_engine(self, input: Option<String>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The version number of the cache engine to be used for the clusters in this replication group. To view the supported cache engine versions, use the DescribeCacheEngineVersions operation.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version) in the ElastiCache User Guide, but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cluster or replication group and create it anew with the earlier engine version.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn cache_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the parameter group to associate with this replication group. If this argument is omitted, the default cache parameter group for the specified engine is used.

If you are running Redis version 3.2.4 or later, only one node group (shard), and want to use a default parameter group, we recommend that you specify the parameter group by name.

To create a Redis (cluster mode disabled) replication group, use CacheParameterGroupName=default.redis3.2.

To create a Redis (cluster mode enabled) replication group, use CacheParameterGroupName=default.redis3.2.cluster.on.

pub fn set_cache_parameter_group_name(self, input: Option<String>) -> Self

pub fn cache_subnet_group_name(self, input: impl Into<String>) -> Self

The name of the cache subnet group to be used for the replication group.

If you're going to launch your cluster in an Amazon VPC, you need to create a subnet group before you start creating a cluster. For more information, see Subnets and Subnet Groups.

pub fn set_cache_subnet_group_name(self, input: Option<String>) -> Self

pub fn cache_security_group_names(self, input: impl Into<String>) -> Self

pub fn set_cache_security_group_names(self, input: Option<Vec<String>>) -> Self

pub fn security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn snapshot_arns(self, input: impl Into<String>) -> Self

pub fn set_snapshot_arns(self, input: Option<Vec<String>>) -> Self

pub fn snapshot_name(self, input: impl Into<String>) -> Self

The name of a snapshot from which to restore data into the new replication group. The snapshot status changes to restoring while the new replication group is being created.

pub fn set_snapshot_name(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port number on which each member of the replication group accepts connections.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn notification_topic_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

The Amazon SNS topic owner must be the same as the cluster owner.

pub fn set_notification_topic_arn(self, input: Option<String>) -> Self

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

This parameter is currently disabled.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn snapshot_retention_limit(self, input: i32) -> Self

The number of days for which ElastiCache retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

Default: 0 (i.e., automatic backups are disabled for this cluster).

pub fn set_snapshot_retention_limit(self, input: Option<i32>) -> Self

pub fn snapshot_window(self, input: impl Into<String>) -> Self

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard).

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

pub fn set_snapshot_window(self, input: Option<String>) -> Self

pub fn auth_token(self, input: impl Into<String>) -> Self

Reserved parameter. The password used to access a password protected server.

AuthToken can be specified only on replication groups where TransitEncryptionEnabled is true.

For HIPAA compliance, you must specify TransitEncryptionEnabled as true, an AuthToken, and a CacheSubnetGroup.

Password constraints:

Must be only printable ASCII characters.

Must be at least 16 characters and no more than 128 characters in length.

The only permitted printable special characters are !, &, #, $, ^, <, >, and -. Other printable special characters cannot be used in the AUTH token.

For more information, see AUTH password at http://redis.io/commands/AUTH.

pub fn set_auth_token(self, input: Option<String>) -> Self

pub fn transit_encryption_enabled(self, input: bool) -> Self

A flag that enables in-transit encryption when set to true.

You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

This parameter is valid only if the Engine parameter is redis, the EngineVersion parameter is 3.2.6, 4.x or later, and the cluster is being created in an Amazon VPC.

If you enable in-transit encryption, you must also specify a value for CacheSubnetGroup.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

For HIPAA compliance, you must specify TransitEncryptionEnabled as true, an AuthToken, and a CacheSubnetGroup.

pub fn set_transit_encryption_enabled(self, input: Option<bool>) -> Self

pub fn at_rest_encryption_enabled(self, input: bool) -> Self

A flag that enables encryption at rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the replication group is created. To enable encryption at rest on a replication group you must set AtRestEncryptionEnabled to true when you create the replication group.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

pub fn set_at_rest_encryption_enabled(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The ID of the KMS key used to encrypt the disk in the cluster.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn user_group_ids(self, input: impl Into<String>) -> Self

pub fn set_user_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn log_delivery_configurations(
self,
input: impl Into<LogDeliveryConfigurationRequest>
) -> Self

pub fn set_log_delivery_configurations(
self,
input: Option<Vec<LogDeliveryConfigurationRequest>>
) -> Self

pub fn build(self) -> Result<CreateReplicationGroupInput, BuildError>

Consumes the builder and constructs a CreateReplicationGroupInput