Struct aws_sdk_elasticache::input::modify_cache_cluster_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ModifyCacheClusterInput

Implementations

impl Builder

pub fn cache_cluster_id(self, input: impl Into<String>) -> Self

The cluster identifier. This value is stored as a lowercase string.

pub fn set_cache_cluster_id(self, input: Option<String>) -> Self

pub fn num_cache_nodes(self, input: i32) -> Self

The number of cache nodes that the cluster should have. If the value for NumCacheNodes is greater than the sum of the number of current cache nodes and the number of cache nodes pending creation (which may be zero), more nodes are added. If the value is less than the number of existing cache nodes, nodes are removed. If the value is equal to the number of current cache nodes, any pending add or remove requests are canceled.

If you are removing cache nodes, you must use the CacheNodeIdsToRemove parameter to provide the IDs of the specific cache nodes to remove.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 40.

Adding or removing Memcached cache nodes can be applied immediately or as a pending operation (see ApplyImmediately).

A pending operation to modify the number of cache nodes in a cluster during its maintenance window, whether by adding or removing nodes in accordance with the scale out architecture, is not queued. The customer's latest request to add or remove nodes to the cluster overrides any previous pending operations to modify the number of cache nodes in the cluster. For example, a request to remove 2 nodes would override a previous pending operation to remove 3 nodes. Similarly, a request to add 2 nodes would override a previous pending operation to remove 3 nodes and vice versa. As Memcached cache nodes may now be provisioned in different Availability Zones with flexible cache node placement, a request to add nodes does not automatically override a previous pending operation to add nodes. The customer can modify the previous pending operation to add more nodes or explicitly cancel the pending request and retry the new request. To cancel pending operations to modify the number of cache nodes in a cluster, use the ModifyCacheCluster request and set NumCacheNodes equal to the number of cache nodes currently in the cluster.

pub fn set_num_cache_nodes(self, input: Option<i32>) -> Self

pub fn cache_node_ids_to_remove(self, input: impl Into<String>) -> Self

pub fn set_cache_node_ids_to_remove(self, input: Option<Vec<String>>) -> Self

pub fn az_mode(self, input: AzMode) -> Self

Specifies whether the new nodes in this Memcached cluster are all created in a single Availability Zone or created across multiple Availability Zones.

Valid values: single-az | cross-az.

This option is only supported for Memcached clusters.

You cannot specify single-az if the Memcached cluster already has cache nodes in different Availability Zones. If cross-az is specified, existing Memcached nodes remain in their current Availability Zone.

Only newly created nodes are located in different Availability Zones.

pub fn set_az_mode(self, input: Option<AzMode>) -> Self

pub fn new_availability_zones(self, input: impl Into<String>) -> Self

pub fn set_new_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn cache_security_group_names(self, input: impl Into<String>) -> Self

pub fn set_cache_security_group_names(self, input: Option<Vec<String>>) -> Self

pub fn security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn notification_topic_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

The Amazon SNS topic owner must be same as the cluster owner.

pub fn set_notification_topic_arn(self, input: Option<String>) -> Self

pub fn cache_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the cache parameter group to apply to this cluster. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

pub fn set_cache_parameter_group_name(self, input: Option<String>) -> Self

pub fn notification_topic_status(self, input: impl Into<String>) -> Self

The status of the Amazon SNS notification topic. Notifications are sent only if the status is active.

Valid values: active | inactive

pub fn set_notification_topic_status(self, input: Option<String>) -> Self

pub fn apply_immediately(self, input: bool) -> Self

If true, this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the cluster.

If false, changes to the cluster are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first.

If you perform a ModifyCacheCluster before a pending modification is applied, the pending modification is replaced by the newer modification.

Valid values: true | false

Default: false

pub fn set_apply_immediately(self, input: Option<bool>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The upgraded version of the cache engine to be run on the cache nodes.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cluster and create it anew with the earlier engine version.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

This parameter is currently disabled.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn snapshot_retention_limit(self, input: i32) -> Self

The number of days for which ElastiCache retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

pub fn set_snapshot_retention_limit(self, input: Option<i32>) -> Self

pub fn snapshot_window(self, input: impl Into<String>) -> Self

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your cluster.

pub fn set_snapshot_window(self, input: Option<String>) -> Self

pub fn cache_node_type(self, input: impl Into<String>) -> Self

A valid cache node type that you want to scale this cluster up to.

pub fn set_cache_node_type(self, input: Option<String>) -> Self

pub fn auth_token(self, input: impl Into<String>) -> Self

Reserved parameter. The password used to access a password protected server. This parameter must be specified with the auth-token-update parameter. Password constraints:

Must be only printable ASCII characters

Must be at least 16 characters and no more than 128 characters in length

Cannot contain any of the following characters: '/', '"', or '@', '%'

For more information, see AUTH password at AUTH.

pub fn set_auth_token(self, input: Option<String>) -> Self

pub fn auth_token_update_strategy(
self,
input: AuthTokenUpdateStrategyType
) -> Self

Specifies the strategy to use to update the AUTH token. This parameter must be specified with the auth-token parameter. Possible values:

Rotate

Set

For more information, see Authenticating Users with Redis AUTH

pub fn set_auth_token_update_strategy(
self,
input: Option<AuthTokenUpdateStrategyType>
) -> Self

pub fn log_delivery_configurations(
self,
input: impl Into<LogDeliveryConfigurationRequest>
) -> Self

pub fn set_log_delivery_configurations(
self,
input: Option<Vec<LogDeliveryConfigurationRequest>>
) -> Self

pub fn build(self) -> Result<ModifyCacheClusterInput, BuildError>

Consumes the builder and constructs a ModifyCacheClusterInput