Struct aws_sdk_elasticache::input::modify_replication_group_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ModifyReplicationGroupInput

Implementations

impl Builder

pub fn replication_group_id(self, input: impl Into<String>) -> Self

The identifier of the replication group to modify.

pub fn set_replication_group_id(self, input: Option<String>) -> Self

pub fn replication_group_description(self, input: impl Into<String>) -> Self

A description for the replication group. Maximum length is 255 characters.

pub fn set_replication_group_description(self, input: Option<String>) -> Self

pub fn primary_cluster_id(self, input: impl Into<String>) -> Self

For replication groups with a single primary, if this parameter is specified, ElastiCache promotes the specified cluster in the specified replication group to the primary role. The nodes of all other clusters in the replication group are read replicas.

pub fn set_primary_cluster_id(self, input: Option<String>) -> Self

pub fn snapshotting_cluster_id(self, input: impl Into<String>) -> Self

The cluster ID that is used as the daily snapshot source for the replication group. This parameter cannot be set for Redis (cluster mode enabled) replication groups.

pub fn set_snapshotting_cluster_id(self, input: Option<String>) -> Self

pub fn automatic_failover_enabled(self, input: bool) -> Self

Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure.

Valid values: true | false

pub fn set_automatic_failover_enabled(self, input: Option<bool>) -> Self

pub fn multi_az_enabled(self, input: bool) -> Self

A flag to indicate MultiAZ is enabled.

pub fn set_multi_az_enabled(self, input: Option<bool>) -> Self

pub fn node_group_id(self, input: impl Into<String>) -> Self

Deprecated. This parameter is not used.

pub fn set_node_group_id(self, input: Option<String>) -> Self

pub fn cache_security_group_names(self, input: impl Into<String>) -> Self

pub fn set_cache_security_group_names(self, input: Option<Vec<String>>) -> Self

pub fn security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn notification_topic_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

The Amazon SNS topic owner must be same as the replication group owner.

pub fn set_notification_topic_arn(self, input: Option<String>) -> Self

pub fn cache_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the cache parameter group to apply to all of the clusters in this replication group. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

pub fn set_cache_parameter_group_name(self, input: Option<String>) -> Self

pub fn notification_topic_status(self, input: impl Into<String>) -> Self

The status of the Amazon SNS notification topic for the replication group. Notifications are sent only if the status is active.

Valid values: active | inactive

pub fn set_notification_topic_status(self, input: Option<String>) -> Self

pub fn apply_immediately(self, input: bool) -> Self

If true, this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the replication group.

If false, changes to the nodes in the replication group are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first.

Valid values: true | false

Default: false

pub fn set_apply_immediately(self, input: Option<bool>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The upgraded version of the cache engine to be run on the clusters in the replication group.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing replication group and create it anew with the earlier engine version.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

This parameter is currently disabled.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn snapshot_retention_limit(self, input: i32) -> Self

The number of days for which ElastiCache retains automatic node group (shard) snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

pub fn set_snapshot_retention_limit(self, input: Option<i32>) -> Self

pub fn snapshot_window(self, input: impl Into<String>) -> Self

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of the node group (shard) specified by SnapshottingClusterId.

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

pub fn set_snapshot_window(self, input: Option<String>) -> Self

pub fn cache_node_type(self, input: impl Into<String>) -> Self

A valid cache node type that you want to scale this replication group to.

pub fn set_cache_node_type(self, input: Option<String>) -> Self

pub fn auth_token(self, input: impl Into<String>) -> Self

Reserved parameter. The password used to access a password protected server. This parameter must be specified with the auth-token-update-strategy parameter. Password constraints:

Must be only printable ASCII characters

Must be at least 16 characters and no more than 128 characters in length

Cannot contain any of the following characters: '/', '"', or '@', '%'

For more information, see AUTH password at AUTH.

pub fn set_auth_token(self, input: Option<String>) -> Self

pub fn auth_token_update_strategy(
self,
input: AuthTokenUpdateStrategyType
) -> Self

Specifies the strategy to use to update the AUTH token. This parameter must be specified with the auth-token parameter. Possible values:

Rotate

Set

For more information, see Authenticating Users with Redis AUTH

pub fn set_auth_token_update_strategy(
self,
input: Option<AuthTokenUpdateStrategyType>
) -> Self

pub fn user_group_ids_to_add(self, input: impl Into<String>) -> Self

pub fn set_user_group_ids_to_add(self, input: Option<Vec<String>>) -> Self

pub fn user_group_ids_to_remove(self, input: impl Into<String>) -> Self

pub fn set_user_group_ids_to_remove(self, input: Option<Vec<String>>) -> Self

pub fn remove_user_groups(self, input: bool) -> Self

Removes the user group associated with this replication group.

pub fn set_remove_user_groups(self, input: Option<bool>) -> Self

pub fn log_delivery_configurations(
self,
input: impl Into<LogDeliveryConfigurationRequest>
) -> Self

pub fn set_log_delivery_configurations(
self,
input: Option<Vec<LogDeliveryConfigurationRequest>>
) -> Self

pub fn build(self) -> Result<ModifyReplicationGroupInput, BuildError>

Consumes the builder and constructs a ModifyReplicationGroupInput