Struct aws_sdk_elasticache::input::ModifyReplicationGroupInput

#[non_exhaustive]
pub struct ModifyReplicationGroupInput {

    pub replication_group_id: Option<String>,
    pub replication_group_description: Option<String>,
    pub primary_cluster_id: Option<String>,
    pub snapshotting_cluster_id: Option<String>,
    pub automatic_failover_enabled: Option<bool>,
    pub multi_az_enabled: Option<bool>,
    pub node_group_id: Option<String>,
    pub cache_security_group_names: Option<Vec<String>>,
    pub security_group_ids: Option<Vec<String>>,
    pub preferred_maintenance_window: Option<String>,
    pub notification_topic_arn: Option<String>,
    pub cache_parameter_group_name: Option<String>,
    pub notification_topic_status: Option<String>,
    pub apply_immediately: bool,
    pub engine_version: Option<String>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub snapshot_retention_limit: Option<i32>,
    pub snapshot_window: Option<String>,
    pub cache_node_type: Option<String>,
    pub auth_token: Option<String>,
    pub auth_token_update_strategy: Option<AuthTokenUpdateStrategyType>,
    pub user_group_ids_to_add: Option<Vec<String>>,
    pub user_group_ids_to_remove: Option<Vec<String>>,
    pub remove_user_groups: Option<bool>,
    pub log_delivery_configurations: Option<Vec<LogDeliveryConfigurationRequest>>,

}

Represents the input of a ModifyReplicationGroups operation.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

replication_group_id: Option<String>

The identifier of the replication group to modify.

replication_group_description: Option<String>

A description for the replication group. Maximum length is 255 characters.

primary_cluster_id: Option<String>

For replication groups with a single primary, if this parameter is specified, ElastiCache promotes the specified cluster in the specified replication group to the primary role. The nodes of all other clusters in the replication group are read replicas.

snapshotting_cluster_id: Option<String>

The cluster ID that is used as the daily snapshot source for the replication group. This parameter cannot be set for Redis (cluster mode enabled) replication groups.

automatic_failover_enabled: Option<bool>

Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure.

Valid values: true | false

multi_az_enabled: Option<bool>

A flag to indicate MultiAZ is enabled.

node_group_id: Option<String>

Deprecated. This parameter is not used.

cache_security_group_names: Option<Vec<String>>

A list of cache security group names to authorize for the clusters in this replication group. This change is asynchronously applied as soon as possible.

This parameter can be used only with replication group containing clusters running outside of an Amazon Virtual Private Cloud (Amazon VPC).

Constraints: Must contain no more than 255 alphanumeric characters. Must not be Default.

security_group_ids: Option<Vec<String>>

Specifies the VPC Security Groups associated with the clusters in the replication group.

This parameter can be used only with replication group containing clusters running in an Amazon Virtual Private Cloud (Amazon VPC).

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

notification_topic_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

The Amazon SNS topic owner must be same as the replication group owner.

cache_parameter_group_name: Option<String>

The name of the cache parameter group to apply to all of the clusters in this replication group. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

notification_topic_status: Option<String>

The status of the Amazon SNS notification topic for the replication group. Notifications are sent only if the status is active.

Valid values: active | inactive

apply_immediately: bool

If true, this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the replication group.

If false, changes to the nodes in the replication group are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first.

Valid values: true | false

Default: false

engine_version: Option<String>

The upgraded version of the cache engine to be run on the clusters in the replication group.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing replication group and create it anew with the earlier engine version.

auto_minor_version_upgrade: Option<bool>

This parameter is currently disabled.

snapshot_retention_limit: Option<i32>

The number of days for which ElastiCache retains automatic node group (shard) snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

snapshot_window: Option<String>

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of the node group (shard) specified by SnapshottingClusterId.

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

cache_node_type: Option<String>

A valid cache node type that you want to scale this replication group to.

auth_token: Option<String>

Reserved parameter. The password used to access a password protected server. This parameter must be specified with the auth-token-update-strategy parameter. Password constraints:

Must be only printable ASCII characters

Must be at least 16 characters and no more than 128 characters in length

Cannot contain any of the following characters: '/', '"', or '@', '%'

For more information, see AUTH password at AUTH.

auth_token_update_strategy: Option<AuthTokenUpdateStrategyType>

Specifies the strategy to use to update the AUTH token. This parameter must be specified with the auth-token parameter. Possible values:

Rotate

Set

For more information, see Authenticating Users with Redis AUTH

user_group_ids_to_add: Option<Vec<String>>

The ID of the user group you are associating with the replication group.

user_group_ids_to_remove: Option<Vec<String>>

The ID of the user group to disassociate from the replication group, meaning the users in the group no longer can access the replication group.

remove_user_groups: Option<bool>

Removes the user group associated with this replication group.

log_delivery_configurations: Option<Vec<LogDeliveryConfigurationRequest>>

Specifies the destination, format and type of the logs.

Implementations

impl ModifyReplicationGroupInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<ModifyReplicationGroup, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<ModifyReplicationGroup>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ModifyReplicationGroupInput