Module aws_sdk_elasticache::model

Modules

authentication

See Authentication

availability_zone

See AvailabilityZone

cache_cluster

See CacheCluster

cache_engine_version

See CacheEngineVersion

cache_node

See CacheNode

cache_node_type_specific_parameter

See CacheNodeTypeSpecificParameter

cache_node_type_specific_value

See CacheNodeTypeSpecificValue

cache_node_update_status

See CacheNodeUpdateStatus

cache_parameter_group

See CacheParameterGroup

cache_parameter_group_status

See CacheParameterGroupStatus

cache_security_group

See CacheSecurityGroup

cache_security_group_membership

See CacheSecurityGroupMembership

cache_subnet_group

See CacheSubnetGroup

cloud_watch_logs_destination_details

See CloudWatchLogsDestinationDetails

configure_shard

See ConfigureShard

customer_node_endpoint

See CustomerNodeEndpoint

destination_details

See DestinationDetails

ec2_security_group

See Ec2SecurityGroup

endpoint

See Endpoint

engine_defaults

See EngineDefaults

event

See Event

filter

See Filter

global_node_group

See GlobalNodeGroup

global_replication_group

See GlobalReplicationGroup

global_replication_group_info

See GlobalReplicationGroupInfo

global_replication_group_member

See GlobalReplicationGroupMember

kinesis_firehose_destination_details

See KinesisFirehoseDestinationDetails

log_delivery_configuration

See LogDeliveryConfiguration

log_delivery_configuration_request

See LogDeliveryConfigurationRequest

node_group

See NodeGroup

node_group_configuration

See NodeGroupConfiguration

node_group_member

See NodeGroupMember

node_group_member_update_status

See NodeGroupMemberUpdateStatus

node_group_update_status

See NodeGroupUpdateStatus

node_snapshot

See NodeSnapshot

notification_configuration

See NotificationConfiguration

parameter

See Parameter

parameter_name_value

See ParameterNameValue

pending_log_delivery_configuration

See PendingLogDeliveryConfiguration

pending_modified_values

See PendingModifiedValues

processed_update_action

See ProcessedUpdateAction

recurring_charge

See RecurringCharge

regional_configuration

See RegionalConfiguration

replication_group

See ReplicationGroup

replication_group_pending_modified_values

See ReplicationGroupPendingModifiedValues

reserved_cache_node

See ReservedCacheNode

reserved_cache_nodes_offering

See ReservedCacheNodesOffering

resharding_configuration

See ReshardingConfiguration

resharding_status

See ReshardingStatus

security_group_membership

See SecurityGroupMembership

service_update

See ServiceUpdate

slot_migration

See SlotMigration

snapshot

See Snapshot

subnet

See Subnet

subnet_outpost

See SubnetOutpost

tag

See Tag

time_range_filter

See TimeRangeFilter

unprocessed_update_action

See UnprocessedUpdateAction

update_action

See UpdateAction

user

See User

user_group

See UserGroup

user_group_pending_changes

See UserGroupPendingChanges

user_groups_update_status

See UserGroupsUpdateStatus

Structs

Authentication

Indicates whether the user requires a password to authenticate.

AvailabilityZone

Describes an Availability Zone in which the cluster is launched.

CacheCluster

Contains all of the attributes of a specific cluster.

CacheEngineVersion

Provides all of the details about a particular cache engine version.

CacheNode

Represents an individual cache node within a cluster. Each cache node runs its own

CacheNodeTypeSpecificParameter

A parameter that has a different value for each cache node type it is applied to. For

CacheNodeTypeSpecificValue

A value that applies only to a certain cache node type.

CacheNodeUpdateStatus

The status of the service update on the cache node

CacheParameterGroup

Represents the output of a CreateCacheParameterGroup operation.

CacheParameterGroupStatus

Status of the cache parameter group.

CacheSecurityGroup

Represents the output of one of the following operations:

CacheSecurityGroupMembership

Represents a cluster's status within a particular cache security group.

CacheSubnetGroup

Represents the output of one of the following operations:

CloudWatchLogsDestinationDetails

The configuration details of the CloudWatch Logs destination.

ConfigureShard

Node group (shard) configuration options when adding or removing replicas.

CustomerNodeEndpoint

The endpoint from which data should be migrated.

DestinationDetails

Configuration details of either a CloudWatch Logs destination or Kinesis Data Firehose destination.

Ec2SecurityGroup

Provides ownership and status information for an Amazon EC2 security group.

Endpoint

Represents the information required for client programs to connect to a cache node.

EngineDefaults

Represents the output of a DescribeEngineDefaultParameters operation.

Event

Represents a single occurrence of something interesting within the system.

Filter

Used to streamline results of a search based on the property being filtered.

GlobalNodeGroup

Indicates the slot configuration and global identifier for a slice group.

GlobalReplicationGroup

Consists of a primary cluster that accepts writes and an associated secondary cluster that resides in a different Amazon region. The secondary cluster accepts only reads. The primary

GlobalReplicationGroupInfo

The name of the Global datastore and role of this replication group in the Global datastore.

GlobalReplicationGroupMember

A member of a Global datastore. It contains the Replication Group Id, the Amazon region and the role of the replication group.

KinesisFirehoseDestinationDetails

The configuration details of the Kinesis Data Firehose destination.

LogDeliveryConfiguration

Returns the destination, format and type of the logs.

LogDeliveryConfigurationRequest

Specifies the destination, format and type of the logs.

NodeGroup

Represents a collection of cache nodes in a replication group.

NodeGroupConfiguration

Node group (shard) configuration options.

NodeGroupMember

Represents a single node within a node group (shard).

NodeGroupMemberUpdateStatus

The status of the service update on the node group member

NodeGroupUpdateStatus

The status of the service update on the node group

NodeSnapshot

Represents an individual cache node in a snapshot of a cluster.

NotificationConfiguration

Describes a notification topic and its status.

Parameter

Describes an individual setting that controls some aspect of ElastiCache behavior.

ParameterNameValue

Describes a name-value pair that is used to update the value of a parameter.

PendingLogDeliveryConfiguration

The log delivery configurations being modified

PendingModifiedValues

A group of settings that are applied to the cluster in the future,

ProcessedUpdateAction

Update action that has been processed for the corresponding apply/stop request

RecurringCharge

Contains the specific price and frequency of a recurring charges for a reserved cache node,

RegionalConfiguration

A list of the replication groups

ReplicationGroup

Contains all of the attributes of a specific Redis replication group.

ReplicationGroupPendingModifiedValues

The settings to be applied to the Redis replication group,

ReservedCacheNode

Represents the output of a PurchaseReservedCacheNodesOffering operation.

ReservedCacheNodesOffering

Describes all of the attributes of a reserved cache node offering.

ReshardingConfiguration

A list of PreferredAvailabilityZones objects that specifies

ReshardingStatus

The status of an online resharding operation.

SecurityGroupMembership

Represents a single cache security group and its status.

ServiceUpdate

An update that you can apply to your Redis clusters.

SlotMigration

Represents the progress of an online resharding operation.

Snapshot

Represents a copy of an entire Redis cluster as of the time when the snapshot was taken.

Subnet

Represents the subnet associated with a cluster.

SubnetOutpost

The ID of the outpost subnet.

Tag

A tag that can be added to an ElastiCache cluster or replication group.

TimeRangeFilter

Filters update actions from the service updates that are in available status during the time range.

UnprocessedUpdateAction

Update action that has failed to be processed for the corresponding apply/stop request

UpdateAction

The status of the service update for a specific replication group

User

UserGroup

UserGroupPendingChanges

Returns the updates being applied to the user group.

UserGroupsUpdateStatus

The status of the user group update.

Enums

AuthTokenUpdateStatus

AuthTokenUpdateStrategyType

AuthenticationType

AutomaticFailoverStatus

AzMode

ChangeType

DestinationType

LogDeliveryConfigurationStatus

LogFormat

LogType

MultiAzStatus

NodeUpdateInitiatedBy

NodeUpdateStatus

OutpostMode

PendingAutomaticFailoverStatus

ServiceUpdateSeverity

ServiceUpdateStatus

ServiceUpdateType

SlaMet

SourceType

UpdateActionStatus