Struct aws_sdk_elasticache::model::cache_cluster::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CacheCluster

Implementations

impl Builder

pub fn cache_cluster_id(self, input: impl Into<String>) -> Self

The user-supplied identifier of the cluster. This identifier is a unique key that identifies a cluster.

pub fn set_cache_cluster_id(self, input: Option<String>) -> Self

pub fn configuration_endpoint(self, input: Endpoint) -> Self

Represents a Memcached cluster endpoint which can be used by an application to connect to any node in the cluster. The configuration endpoint will always have .cfg in it.

Example: mem-3.9dvc4r.cfg.usw2.cache.amazonaws.com:11211

pub fn set_configuration_endpoint(self, input: Option<Endpoint>) -> Self

pub fn client_download_landing_page(self, input: impl Into<String>) -> Self

The URL of the web page where you can download the latest ElastiCache client library.

pub fn set_client_download_landing_page(self, input: Option<String>) -> Self

pub fn cache_node_type(self, input: impl Into<String>) -> Self

The name of the compute and memory capacity node type for the cluster.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

pub fn set_cache_node_type(self, input: Option<String>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

The name of the cache engine (memcached or redis) to be used for this cluster.

pub fn set_engine(self, input: Option<String>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The version of the cache engine that is used in this cluster.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn cache_cluster_status(self, input: impl Into<String>) -> Self

The current state of this cluster, one of the following values: available, creating, deleted, deleting, incompatible-network, modifying, rebooting cluster nodes, restore-failed, or snapshotting.

pub fn set_cache_cluster_status(self, input: Option<String>) -> Self

pub fn num_cache_nodes(self, input: i32) -> Self

The number of cache nodes in the cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 40.

pub fn set_num_cache_nodes(self, input: Option<i32>) -> Self

pub fn preferred_availability_zone(self, input: impl Into<String>) -> Self

The name of the Availability Zone in which the cluster is located or "Multiple" if the cache nodes are located in different Availability Zones.

pub fn set_preferred_availability_zone(self, input: Option<String>) -> Self

pub fn preferred_outpost_arn(self, input: impl Into<String>) -> Self

The outpost ARN in which the cache cluster is created.

pub fn set_preferred_outpost_arn(self, input: Option<String>) -> Self

pub fn cache_cluster_create_time(self, input: Instant) -> Self

The date and time when the cluster was created.

pub fn set_cache_cluster_create_time(self, input: Option<Instant>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn pending_modified_values(self, input: PendingModifiedValues) -> Self

A group of settings that are applied to the cluster in the future, or that are currently being applied.

pub fn set_pending_modified_values(
self,
input: Option<PendingModifiedValues>
) -> Self

pub fn notification_configuration(
self,
input: NotificationConfiguration
) -> Self

Describes a notification topic and its status. Notification topics are used for publishing ElastiCache events to subscribers using Amazon Simple Notification Service (SNS).

pub fn set_notification_configuration(
self,
input: Option<NotificationConfiguration>
) -> Self

pub fn cache_security_groups(
self,
input: impl Into<CacheSecurityGroupMembership>
) -> Self

pub fn set_cache_security_groups(
self,
input: Option<Vec<CacheSecurityGroupMembership>>
) -> Self

pub fn cache_parameter_group(self, input: CacheParameterGroupStatus) -> Self

Status of the cache parameter group.

pub fn set_cache_parameter_group(
self,
input: Option<CacheParameterGroupStatus>
) -> Self

pub fn cache_subnet_group_name(self, input: impl Into<String>) -> Self

The name of the cache subnet group associated with the cluster.

pub fn set_cache_subnet_group_name(self, input: Option<String>) -> Self

pub fn cache_nodes(self, input: impl Into<CacheNode>) -> Self

pub fn set_cache_nodes(self, input: Option<Vec<CacheNode>>) -> Self

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

This parameter is currently disabled.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn security_groups(self, input: impl Into<SecurityGroupMembership>) -> Self

pub fn set_security_groups(
self,
input: Option<Vec<SecurityGroupMembership>>
) -> Self

pub fn replication_group_id(self, input: impl Into<String>) -> Self

The replication group to which this cluster belongs. If this field is empty, the cluster is not associated with any replication group.

pub fn set_replication_group_id(self, input: Option<String>) -> Self

pub fn snapshot_retention_limit(self, input: i32) -> Self

The number of days for which ElastiCache retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

pub fn set_snapshot_retention_limit(self, input: Option<i32>) -> Self

pub fn snapshot_window(self, input: impl Into<String>) -> Self

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your cluster.

Example: 05:00-09:00

pub fn set_snapshot_window(self, input: Option<String>) -> Self

pub fn auth_token_enabled(self, input: bool) -> Self

A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

pub fn set_auth_token_enabled(self, input: Option<bool>) -> Self

pub fn auth_token_last_modified_date(self, input: Instant) -> Self

The date the auth token was last modified

pub fn set_auth_token_last_modified_date(self, input: Option<Instant>) -> Self

pub fn transit_encryption_enabled(self, input: bool) -> Self

A flag that enables in-transit encryption when set to true.

You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

pub fn set_transit_encryption_enabled(self, input: Option<bool>) -> Self

pub fn at_rest_encryption_enabled(self, input: bool) -> Self

A flag that enables encryption at-rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the cluster is created. To enable at-rest encryption on a cluster you must set AtRestEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

pub fn set_at_rest_encryption_enabled(self, input: Option<bool>) -> Self

pub fn arn(self, input: impl Into<String>) -> Self

The ARN (Amazon Resource Name) of the cache cluster.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn replication_group_log_delivery_enabled(self, input: bool) -> Self

A boolean value indicating whether log delivery is enabled for the replication group.

pub fn set_replication_group_log_delivery_enabled(
self,
input: Option<bool>
) -> Self

pub fn log_delivery_configurations(
self,
input: impl Into<LogDeliveryConfiguration>
) -> Self

pub fn set_log_delivery_configurations(
self,
input: Option<Vec<LogDeliveryConfiguration>>
) -> Self

pub fn build(self) -> CacheCluster

Consumes the builder and constructs a CacheCluster