Struct aws_sdk_elasticache::model::replication_group::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ReplicationGroup

Implementations

impl Builder

pub fn replication_group_id(self, input: impl Into<String>) -> Self

The identifier for the replication group.

pub fn set_replication_group_id(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The user supplied description of the replication group.

pub fn set_description(self, input: Option<String>) -> Self

pub fn global_replication_group_info(
self,
input: GlobalReplicationGroupInfo
) -> Self

The name of the Global datastore and role of this replication group in the Global datastore.

pub fn set_global_replication_group_info(
self,
input: Option<GlobalReplicationGroupInfo>
) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The current state of this replication group - creating, available, modifying, deleting, create-failed, snapshotting.

pub fn set_status(self, input: Option<String>) -> Self

pub fn pending_modified_values(
self,
input: ReplicationGroupPendingModifiedValues
) -> Self

A group of settings to be applied to the replication group, either immediately or during the next maintenance window.

pub fn set_pending_modified_values(
self,
input: Option<ReplicationGroupPendingModifiedValues>
) -> Self

pub fn member_clusters(self, input: impl Into<String>) -> Self

pub fn set_member_clusters(self, input: Option<Vec<String>>) -> Self

pub fn node_groups(self, input: impl Into<NodeGroup>) -> Self

pub fn set_node_groups(self, input: Option<Vec<NodeGroup>>) -> Self

pub fn snapshotting_cluster_id(self, input: impl Into<String>) -> Self

The cluster ID that is used as the daily snapshot source for the replication group.

pub fn set_snapshotting_cluster_id(self, input: Option<String>) -> Self

pub fn automatic_failover(self, input: AutomaticFailoverStatus) -> Self

Indicates the status of automatic failover for this Redis replication group.

pub fn set_automatic_failover(
self,
input: Option<AutomaticFailoverStatus>
) -> Self

pub fn multi_az(self, input: MultiAzStatus) -> Self

A flag indicating if you have Multi-AZ enabled to enhance fault tolerance. For more information, see Minimizing Downtime: Multi-AZ

pub fn set_multi_az(self, input: Option<MultiAzStatus>) -> Self

pub fn configuration_endpoint(self, input: Endpoint) -> Self

The configuration endpoint for this replication group. Use the configuration endpoint to connect to this replication group.

pub fn set_configuration_endpoint(self, input: Option<Endpoint>) -> Self

pub fn snapshot_retention_limit(self, input: i32) -> Self

The number of days for which ElastiCache retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

pub fn set_snapshot_retention_limit(self, input: Option<i32>) -> Self

pub fn snapshot_window(self, input: impl Into<String>) -> Self

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard).

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

This parameter is only valid if the Engine parameter is redis.

pub fn set_snapshot_window(self, input: Option<String>) -> Self

pub fn cluster_enabled(self, input: bool) -> Self

A flag indicating whether or not this replication group is cluster enabled; i.e., whether its data can be partitioned across multiple shards (API/CLI: node groups).

Valid values: true | false

pub fn set_cluster_enabled(self, input: Option<bool>) -> Self

pub fn cache_node_type(self, input: impl Into<String>) -> Self

The name of the compute and memory capacity node type for each node in the replication group.

pub fn set_cache_node_type(self, input: Option<String>) -> Self

pub fn auth_token_enabled(self, input: bool) -> Self

A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

pub fn set_auth_token_enabled(self, input: Option<bool>) -> Self

pub fn auth_token_last_modified_date(self, input: Instant) -> Self

The date the auth token was last modified

pub fn set_auth_token_last_modified_date(self, input: Option<Instant>) -> Self

pub fn transit_encryption_enabled(self, input: bool) -> Self

A flag that enables in-transit encryption when set to true.

You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

pub fn set_transit_encryption_enabled(self, input: Option<bool>) -> Self

pub fn at_rest_encryption_enabled(self, input: bool) -> Self

A flag that enables encryption at-rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the cluster is created. To enable encryption at-rest on a cluster you must set AtRestEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

pub fn set_at_rest_encryption_enabled(self, input: Option<bool>) -> Self

pub fn member_clusters_outpost_arns(self, input: impl Into<String>) -> Self

pub fn set_member_clusters_outpost_arns(
self,
input: Option<Vec<String>>
) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The ID of the KMS key used to encrypt the disk in the cluster.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn arn(self, input: impl Into<String>) -> Self

The ARN (Amazon Resource Name) of the replication group.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn user_group_ids(self, input: impl Into<String>) -> Self

pub fn set_user_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn log_delivery_configurations(
self,
input: impl Into<LogDeliveryConfiguration>
) -> Self

pub fn set_log_delivery_configurations(
self,
input: Option<Vec<LogDeliveryConfiguration>>
) -> Self

pub fn replication_group_create_time(self, input: Instant) -> Self

The date and time when the cluster was created.

pub fn set_replication_group_create_time(self, input: Option<Instant>) -> Self

pub fn build(self) -> ReplicationGroup

Consumes the builder and constructs a ReplicationGroup