Struct aws_sdk_elasticache::model::Snapshot

#[non_exhaustive]
pub struct Snapshot {

    pub snapshot_name: Option<String>,
    pub replication_group_id: Option<String>,
    pub replication_group_description: Option<String>,
    pub cache_cluster_id: Option<String>,
    pub snapshot_status: Option<String>,
    pub snapshot_source: Option<String>,
    pub cache_node_type: Option<String>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub num_cache_nodes: Option<i32>,
    pub preferred_availability_zone: Option<String>,
    pub preferred_outpost_arn: Option<String>,
    pub cache_cluster_create_time: Option<Instant>,
    pub preferred_maintenance_window: Option<String>,
    pub topic_arn: Option<String>,
    pub port: Option<i32>,
    pub cache_parameter_group_name: Option<String>,
    pub cache_subnet_group_name: Option<String>,
    pub vpc_id: Option<String>,
    pub auto_minor_version_upgrade: bool,
    pub snapshot_retention_limit: Option<i32>,
    pub snapshot_window: Option<String>,
    pub num_node_groups: Option<i32>,
    pub automatic_failover: Option<AutomaticFailoverStatus>,
    pub node_snapshots: Option<Vec<NodeSnapshot>>,
    pub kms_key_id: Option<String>,
    pub arn: Option<String>,

}

Represents a copy of an entire Redis cluster as of the time when the snapshot was taken.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

snapshot_name: Option<String>

The name of a snapshot. For an automatic snapshot, the name is system-generated. For a manual snapshot, this is the user-provided name.

replication_group_id: Option<String>

The unique identifier of the source replication group.

replication_group_description: Option<String>

A description of the source replication group.

cache_cluster_id: Option<String>

The user-supplied identifier of the source cluster.

snapshot_status: Option<String>

The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

snapshot_source: Option<String>

Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

cache_node_type: Option<String>

The name of the compute and memory capacity node type for the source cluster.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

engine: Option<String>

The name of the cache engine (memcached or redis) used by the source cluster.

engine_version: Option<String>

The version of the cache engine version that is used by the source cluster.

num_cache_nodes: Option<i32>

The number of cache nodes in the source cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 40.

preferred_availability_zone: Option<String>

The name of the Availability Zone in which the source cluster is located.

preferred_outpost_arn: Option<String>

The ARN (Amazon Resource Name) of the preferred outpost.

cache_cluster_create_time: Option<Instant>

The date and time when the source cluster was created.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

topic_arn: Option<String>

The Amazon Resource Name (ARN) for the topic used by the source cluster for publishing notifications.

port: Option<i32>

The port number used by each cache nodes in the source cluster.

cache_parameter_group_name: Option<String>

The cache parameter group that is associated with the source cluster.

cache_subnet_group_name: Option<String>

The name of the cache subnet group associated with the source cluster.

vpc_id: Option<String>

The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group for the source cluster.

auto_minor_version_upgrade: bool

This parameter is currently disabled.

snapshot_retention_limit: Option<i32>

For an automatic snapshot, the number of days for which ElastiCache retains the snapshot before deleting it.

For manual snapshots, this field reflects the SnapshotRetentionLimit for the source cluster when the snapshot was created. This field is otherwise ignored: Manual snapshots do not expire, and can only be deleted using the DeleteSnapshot operation.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

snapshot_window: Option<String>

The daily time range during which ElastiCache takes daily snapshots of the source cluster.

num_node_groups: Option<i32>

The number of node groups (shards) in this snapshot. When restoring from a snapshot, the number of node groups (shards) in the snapshot and in the restored replication group must be the same.

automatic_failover: Option<AutomaticFailoverStatus>

Indicates the status of automatic failover for the source Redis replication group.

node_snapshots: Option<Vec<NodeSnapshot>>

A list of the cache nodes in the source cluster.

kms_key_id: Option<String>

The ID of the KMS key used to encrypt the snapshot.

arn: Option<String>

The ARN (Amazon Resource Name) of the snapshot.

Implementations

impl Snapshot

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Snapshot