Module aws_sdk_elasticache::operation

Structs

AddTagsToResource

A tag is a key-value pair where the key and value are case-sensitive.

AuthorizeCacheSecurityGroupIngress

Allows network ingress to a cache

BatchApplyUpdateAction

Apply the service update. For more information on service updates and applying them, see Applying Service Updates.

BatchStopUpdateAction

Stop the service update. For more information on service updates and stopping them, see Stopping Service Updates.

CompleteMigration

Complete the migration of data.

CopySnapshot

Makes a copy of an existing snapshot.

CreateCacheCluster

Creates a cluster. All nodes in the

CreateCacheParameterGroup

Creates a new Amazon ElastiCache cache parameter group. An ElastiCache

CreateCacheSecurityGroup

Creates a new cache security group. Use a

CreateCacheSubnetGroup

Creates a new cache subnet group.

CreateGlobalReplicationGroup

Global Datastore for Redis offers fully managed, fast,

CreateReplicationGroup

Creates a Redis (cluster mode disabled) or a Redis (cluster mode enabled) replication group.

CreateSnapshot

Creates a copy of an entire cluster or replication group at a

CreateUser

For Redis engine version 6.x onwards: Creates a Redis user. For more information, see Using Role Based Access Control (RBAC).

CreateUserGroup

For Redis engine version 6.x onwards: Creates a Redis user group. For more information, see Using Role Based Access Control (RBAC)

DecreaseNodeGroupsInGlobalReplicationGroup

Decreases the number of node groups in a Global datastore

DecreaseReplicaCount

Dynamically decreases the number of replicas in a Redis (cluster mode disabled) replication group or the number of

DeleteCacheCluster

Deletes a previously provisioned cluster.

DeleteCacheParameterGroup

Deletes the specified cache parameter

DeleteCacheSecurityGroup

Deletes a cache security group.

DeleteCacheSubnetGroup

Deletes a cache subnet group.

DeleteGlobalReplicationGroup

Deleting a Global datastore is a two-step process:

DeleteReplicationGroup

Deletes an existing replication group.

DeleteSnapshot

Deletes an existing snapshot. When you receive a

DeleteUser

For Redis engine version 6.x onwards: Deletes a user. The user will be removed from all user groups and in turn removed from all replication groups. For more information, see Using Role Based Access Control (RBAC).

DeleteUserGroup

For Redis engine version 6.x onwards: Deletes a user group. The user group must first be disassociated from the replication group before it can be deleted. For more information, see Using Role Based Access Control (RBAC).

DescribeCacheClusters

Returns information about all provisioned

DescribeCacheEngineVersions

Returns a list of the available cache

DescribeCacheParameterGroups

Returns a list of cache parameter group

DescribeCacheParameters

Returns the detailed parameter list for a

DescribeCacheSecurityGroups

Returns a list of cache security group

DescribeCacheSubnetGroups

Returns a list of cache subnet group

DescribeEngineDefaultParameters

Returns the default engine and

DescribeEvents

Returns events related to clusters, cache

DescribeGlobalReplicationGroups

Returns information about a particular global replication group. If no identifier is specified, returns information about all Global datastores.

DescribeReplicationGroups

Returns information about a particular

DescribeReservedCacheNodes

Returns information about reserved cache

DescribeReservedCacheNodesOfferings

Lists available reserved cache

DescribeServiceUpdates

Returns details of the service updates

DescribeSnapshots

Returns information about cluster or replication group snapshots.

DescribeUpdateActions

Returns details of the update actions

DescribeUserGroups

Returns a list of user groups.

DescribeUsers

Returns a list of users.

DisassociateGlobalReplicationGroup

Remove a secondary cluster from the Global datastore using the Global datastore name. The secondary cluster will no longer receive updates from the primary cluster, but will remain as a standalone cluster in that Amazon region.

FailoverGlobalReplicationGroup

Used to failover the primary region to a selected secondary region. The selected secondary region will become primary, and all other clusters will become secondary.

IncreaseNodeGroupsInGlobalReplicationGroup

Increase the number of node groups in the Global datastore

IncreaseReplicaCount

Dynamically increases the number of replicas in a Redis (cluster mode disabled) replication group or the number of

ListAllowedNodeTypeModifications

Lists all available node types that you

ListTagsForResource

Lists all tags currently on a named resource.

ModifyCacheCluster

Modifies the settings for a cluster. You

ModifyCacheParameterGroup

Modifies the parameters of a cache

ModifyCacheSubnetGroup

Modifies an existing cache subnet group.

ModifyGlobalReplicationGroup

Modifies the settings for a Global datastore.

ModifyReplicationGroup

Modifies the settings for a replication group.

ModifyReplicationGroupShardConfiguration

Modifies a replication group's shards (node groups) by allowing you to

ModifyUser

Changes user password(s) and/or access string.

ModifyUserGroup

Changes the list of users that belong to the user group.

PurchaseReservedCacheNodesOffering

Allows you to purchase a reserved

RebalanceSlotsInGlobalReplicationGroup

Redistribute slots to ensure uniform distribution across existing shards in the cluster.

RebootCacheCluster

Reboots some, or all, of the cache nodes

RemoveTagsFromResource

Removes the tags identified by the TagKeys

ResetCacheParameterGroup

Modifies the parameters of a cache

RevokeCacheSecurityGroupIngress

Revokes ingress from a cache security group.

StartMigration

Start the migration of data.

TestFailover

Represents the input of a TestFailover operation which test automatic failover on