Module aws_sdk_elasticache::output

Modules

add_tags_to_resource_output

See AddTagsToResourceOutput

authorize_cache_security_group_ingress_output

See AuthorizeCacheSecurityGroupIngressOutput

batch_apply_update_action_output

See BatchApplyUpdateActionOutput

batch_stop_update_action_output

See BatchStopUpdateActionOutput

complete_migration_output

See CompleteMigrationOutput

copy_snapshot_output

See CopySnapshotOutput

create_cache_cluster_output

See CreateCacheClusterOutput

create_cache_parameter_group_output

See CreateCacheParameterGroupOutput

create_cache_security_group_output

See CreateCacheSecurityGroupOutput

create_cache_subnet_group_output

See CreateCacheSubnetGroupOutput

create_global_replication_group_output

See CreateGlobalReplicationGroupOutput

create_replication_group_output

See CreateReplicationGroupOutput

create_snapshot_output

See CreateSnapshotOutput

create_user_group_output

See CreateUserGroupOutput

create_user_output

See CreateUserOutput

decrease_node_groups_in_global_replication_group_output

See DecreaseNodeGroupsInGlobalReplicationGroupOutput

decrease_replica_count_output

See DecreaseReplicaCountOutput

delete_cache_cluster_output

See DeleteCacheClusterOutput

delete_cache_parameter_group_output

See DeleteCacheParameterGroupOutput

delete_cache_security_group_output

See DeleteCacheSecurityGroupOutput

delete_cache_subnet_group_output

See DeleteCacheSubnetGroupOutput

delete_global_replication_group_output

See DeleteGlobalReplicationGroupOutput

delete_replication_group_output

See DeleteReplicationGroupOutput

delete_snapshot_output

See DeleteSnapshotOutput

delete_user_group_output

See DeleteUserGroupOutput

delete_user_output

See DeleteUserOutput

describe_cache_clusters_output

See DescribeCacheClustersOutput

describe_cache_engine_versions_output

See DescribeCacheEngineVersionsOutput

describe_cache_parameter_groups_output

See DescribeCacheParameterGroupsOutput

describe_cache_parameters_output

See DescribeCacheParametersOutput

describe_cache_security_groups_output

See DescribeCacheSecurityGroupsOutput

describe_cache_subnet_groups_output

See DescribeCacheSubnetGroupsOutput

describe_engine_default_parameters_output

See DescribeEngineDefaultParametersOutput

describe_events_output

See DescribeEventsOutput

describe_global_replication_groups_output

See DescribeGlobalReplicationGroupsOutput

describe_replication_groups_output

See DescribeReplicationGroupsOutput

describe_reserved_cache_nodes_offerings_output

See DescribeReservedCacheNodesOfferingsOutput

describe_reserved_cache_nodes_output

See DescribeReservedCacheNodesOutput

describe_service_updates_output

See DescribeServiceUpdatesOutput

describe_snapshots_output

See DescribeSnapshotsOutput

describe_update_actions_output

See DescribeUpdateActionsOutput

describe_user_groups_output

See DescribeUserGroupsOutput

describe_users_output

See DescribeUsersOutput

disassociate_global_replication_group_output

See DisassociateGlobalReplicationGroupOutput

failover_global_replication_group_output

See FailoverGlobalReplicationGroupOutput

increase_node_groups_in_global_replication_group_output

See IncreaseNodeGroupsInGlobalReplicationGroupOutput

increase_replica_count_output

See IncreaseReplicaCountOutput

list_allowed_node_type_modifications_output

See ListAllowedNodeTypeModificationsOutput

list_tags_for_resource_output

See ListTagsForResourceOutput

modify_cache_cluster_output

See ModifyCacheClusterOutput

modify_cache_parameter_group_output

See ModifyCacheParameterGroupOutput

modify_cache_subnet_group_output

See ModifyCacheSubnetGroupOutput

modify_global_replication_group_output

See ModifyGlobalReplicationGroupOutput

modify_replication_group_output

See ModifyReplicationGroupOutput

modify_replication_group_shard_configuration_output

See ModifyReplicationGroupShardConfigurationOutput

modify_user_group_output

See ModifyUserGroupOutput

modify_user_output

See ModifyUserOutput

purchase_reserved_cache_nodes_offering_output

See PurchaseReservedCacheNodesOfferingOutput

rebalance_slots_in_global_replication_group_output

See RebalanceSlotsInGlobalReplicationGroupOutput

reboot_cache_cluster_output

See RebootCacheClusterOutput

remove_tags_from_resource_output

See RemoveTagsFromResourceOutput

reset_cache_parameter_group_output

See ResetCacheParameterGroupOutput

revoke_cache_security_group_ingress_output

See RevokeCacheSecurityGroupIngressOutput

start_migration_output

See StartMigrationOutput

test_failover_output

See TestFailoverOutput

Structs

AddTagsToResourceOutput

Represents the output from the AddTagsToResource,

AuthorizeCacheSecurityGroupIngressOutput

BatchApplyUpdateActionOutput

BatchStopUpdateActionOutput

CompleteMigrationOutput

CopySnapshotOutput

CreateCacheClusterOutput

CreateCacheParameterGroupOutput

CreateCacheSecurityGroupOutput

CreateCacheSubnetGroupOutput

CreateGlobalReplicationGroupOutput

CreateReplicationGroupOutput

CreateSnapshotOutput

CreateUserGroupOutput

CreateUserOutput

DecreaseNodeGroupsInGlobalReplicationGroupOutput

DecreaseReplicaCountOutput

DeleteCacheClusterOutput

DeleteCacheParameterGroupOutput

DeleteCacheSecurityGroupOutput

DeleteCacheSubnetGroupOutput

DeleteGlobalReplicationGroupOutput

DeleteReplicationGroupOutput

DeleteSnapshotOutput

DeleteUserGroupOutput

DeleteUserOutput

DescribeCacheClustersOutput

Represents the output of a DescribeCacheClusters operation.

DescribeCacheEngineVersionsOutput

Represents the output of a DescribeCacheEngineVersions operation.

DescribeCacheParameterGroupsOutput

Represents the output of a DescribeCacheParameterGroups operation.

DescribeCacheParametersOutput

Represents the output of a DescribeCacheParameters operation.

DescribeCacheSecurityGroupsOutput

Represents the output of a DescribeCacheSecurityGroups operation.

DescribeCacheSubnetGroupsOutput

Represents the output of a DescribeCacheSubnetGroups operation.

DescribeEngineDefaultParametersOutput

DescribeEventsOutput

Represents the output of a DescribeEvents operation.

DescribeGlobalReplicationGroupsOutput

DescribeReplicationGroupsOutput

Represents the output of a DescribeReplicationGroups operation.

DescribeReservedCacheNodesOfferingsOutput

Represents the output of a DescribeReservedCacheNodesOfferings operation.

DescribeReservedCacheNodesOutput

Represents the output of a DescribeReservedCacheNodes operation.

DescribeServiceUpdatesOutput

DescribeSnapshotsOutput

Represents the output of a DescribeSnapshots operation.

DescribeUpdateActionsOutput

DescribeUserGroupsOutput

DescribeUsersOutput

DisassociateGlobalReplicationGroupOutput

FailoverGlobalReplicationGroupOutput

IncreaseNodeGroupsInGlobalReplicationGroupOutput

IncreaseReplicaCountOutput

ListAllowedNodeTypeModificationsOutput

Represents the allowed node types you can use to modify your cluster or replication group.

ListTagsForResourceOutput

Represents the output from the AddTagsToResource,

ModifyCacheClusterOutput

ModifyCacheParameterGroupOutput

Represents the output of one of the following operations:

ModifyCacheSubnetGroupOutput

ModifyGlobalReplicationGroupOutput

ModifyReplicationGroupOutput

ModifyReplicationGroupShardConfigurationOutput

ModifyUserGroupOutput

ModifyUserOutput

PurchaseReservedCacheNodesOfferingOutput

RebalanceSlotsInGlobalReplicationGroupOutput

RebootCacheClusterOutput

RemoveTagsFromResourceOutput

Represents the output from the AddTagsToResource,

ResetCacheParameterGroupOutput

Represents the output of one of the following operations:

RevokeCacheSecurityGroupIngressOutput

StartMigrationOutput

TestFailoverOutput