Struct aws_sdk_elasticache::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonElastiCacheV9.

This client allows ergonomic access to a AmazonElastiCacheV9-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_tags_to_resource(&self) -> AddTagsToResource<C, M, R>

Constructs a fluent builder for the AddTagsToResource operation.

See AddTagsToResource for more information about the operation and its arguments.

pub fn authorize_cache_security_group_ingress(
&self
) -> AuthorizeCacheSecurityGroupIngress<C, M, R>

Constructs a fluent builder for the AuthorizeCacheSecurityGroupIngress operation.

See AuthorizeCacheSecurityGroupIngress for more information about the operation and its arguments.

pub fn batch_apply_update_action(&self) -> BatchApplyUpdateAction<C, M, R>

Constructs a fluent builder for the BatchApplyUpdateAction operation.

See BatchApplyUpdateAction for more information about the operation and its arguments.

pub fn batch_stop_update_action(&self) -> BatchStopUpdateAction<C, M, R>

Constructs a fluent builder for the BatchStopUpdateAction operation.

See BatchStopUpdateAction for more information about the operation and its arguments.

pub fn complete_migration(&self) -> CompleteMigration<C, M, R>

Constructs a fluent builder for the CompleteMigration operation.

See CompleteMigration for more information about the operation and its arguments.

pub fn copy_snapshot(&self) -> CopySnapshot<C, M, R>

Constructs a fluent builder for the CopySnapshot operation.

See CopySnapshot for more information about the operation and its arguments.

pub fn create_cache_cluster(&self) -> CreateCacheCluster<C, M, R>

Constructs a fluent builder for the CreateCacheCluster operation.

See CreateCacheCluster for more information about the operation and its arguments.

pub fn create_cache_parameter_group(&self) -> CreateCacheParameterGroup<C, M, R>

Constructs a fluent builder for the CreateCacheParameterGroup operation.

See CreateCacheParameterGroup for more information about the operation and its arguments.

pub fn create_cache_security_group(&self) -> CreateCacheSecurityGroup<C, M, R>

Constructs a fluent builder for the CreateCacheSecurityGroup operation.

See CreateCacheSecurityGroup for more information about the operation and its arguments.

pub fn create_cache_subnet_group(&self) -> CreateCacheSubnetGroup<C, M, R>

Constructs a fluent builder for the CreateCacheSubnetGroup operation.

See CreateCacheSubnetGroup for more information about the operation and its arguments.

pub fn create_global_replication_group(
&self
) -> CreateGlobalReplicationGroup<C, M, R>

Constructs a fluent builder for the CreateGlobalReplicationGroup operation.

See CreateGlobalReplicationGroup for more information about the operation and its arguments.

pub fn create_replication_group(&self) -> CreateReplicationGroup<C, M, R>

Constructs a fluent builder for the CreateReplicationGroup operation.

See CreateReplicationGroup for more information about the operation and its arguments.

pub fn create_snapshot(&self) -> CreateSnapshot<C, M, R>

Constructs a fluent builder for the CreateSnapshot operation.

See CreateSnapshot for more information about the operation and its arguments.

pub fn create_user(&self) -> CreateUser<C, M, R>

Constructs a fluent builder for the CreateUser operation.

See CreateUser for more information about the operation and its arguments.

pub fn create_user_group(&self) -> CreateUserGroup<C, M, R>

Constructs a fluent builder for the CreateUserGroup operation.

See CreateUserGroup for more information about the operation and its arguments.

pub fn decrease_node_groups_in_global_replication_group(
&self
) -> DecreaseNodeGroupsInGlobalReplicationGroup<C, M, R>

Constructs a fluent builder for the DecreaseNodeGroupsInGlobalReplicationGroup operation.

See DecreaseNodeGroupsInGlobalReplicationGroup for more information about the operation and its arguments.

pub fn decrease_replica_count(&self) -> DecreaseReplicaCount<C, M, R>

Constructs a fluent builder for the DecreaseReplicaCount operation.

See DecreaseReplicaCount for more information about the operation and its arguments.

pub fn delete_cache_cluster(&self) -> DeleteCacheCluster<C, M, R>

Constructs a fluent builder for the DeleteCacheCluster operation.

See DeleteCacheCluster for more information about the operation and its arguments.

pub fn delete_cache_parameter_group(&self) -> DeleteCacheParameterGroup<C, M, R>

Constructs a fluent builder for the DeleteCacheParameterGroup operation.

See DeleteCacheParameterGroup for more information about the operation and its arguments.

pub fn delete_cache_security_group(&self) -> DeleteCacheSecurityGroup<C, M, R>

Constructs a fluent builder for the DeleteCacheSecurityGroup operation.

See DeleteCacheSecurityGroup for more information about the operation and its arguments.

pub fn delete_cache_subnet_group(&self) -> DeleteCacheSubnetGroup<C, M, R>

Constructs a fluent builder for the DeleteCacheSubnetGroup operation.

See DeleteCacheSubnetGroup for more information about the operation and its arguments.

pub fn delete_global_replication_group(
&self
) -> DeleteGlobalReplicationGroup<C, M, R>

Constructs a fluent builder for the DeleteGlobalReplicationGroup operation.

See DeleteGlobalReplicationGroup for more information about the operation and its arguments.

pub fn delete_replication_group(&self) -> DeleteReplicationGroup<C, M, R>

Constructs a fluent builder for the DeleteReplicationGroup operation.

See DeleteReplicationGroup for more information about the operation and its arguments.

pub fn delete_snapshot(&self) -> DeleteSnapshot<C, M, R>

Constructs a fluent builder for the DeleteSnapshot operation.

See DeleteSnapshot for more information about the operation and its arguments.

pub fn delete_user(&self) -> DeleteUser<C, M, R>

Constructs a fluent builder for the DeleteUser operation.

See DeleteUser for more information about the operation and its arguments.

pub fn delete_user_group(&self) -> DeleteUserGroup<C, M, R>

Constructs a fluent builder for the DeleteUserGroup operation.

See DeleteUserGroup for more information about the operation and its arguments.

pub fn describe_cache_clusters(&self) -> DescribeCacheClusters<C, M, R>

Constructs a fluent builder for the DescribeCacheClusters operation.

See DescribeCacheClusters for more information about the operation and its arguments.

pub fn describe_cache_engine_versions(
&self
) -> DescribeCacheEngineVersions<C, M, R>

Constructs a fluent builder for the DescribeCacheEngineVersions operation.

See DescribeCacheEngineVersions for more information about the operation and its arguments.

pub fn describe_cache_parameter_groups(
&self
) -> DescribeCacheParameterGroups<C, M, R>

Constructs a fluent builder for the DescribeCacheParameterGroups operation.

See DescribeCacheParameterGroups for more information about the operation and its arguments.

pub fn describe_cache_parameters(&self) -> DescribeCacheParameters<C, M, R>

Constructs a fluent builder for the DescribeCacheParameters operation.

See DescribeCacheParameters for more information about the operation and its arguments.

pub fn describe_cache_security_groups(
&self
) -> DescribeCacheSecurityGroups<C, M, R>

Constructs a fluent builder for the DescribeCacheSecurityGroups operation.

See DescribeCacheSecurityGroups for more information about the operation and its arguments.

pub fn describe_cache_subnet_groups(&self) -> DescribeCacheSubnetGroups<C, M, R>

Constructs a fluent builder for the DescribeCacheSubnetGroups operation.

See DescribeCacheSubnetGroups for more information about the operation and its arguments.

pub fn describe_engine_default_parameters(
&self
) -> DescribeEngineDefaultParameters<C, M, R>

Constructs a fluent builder for the DescribeEngineDefaultParameters operation.

See DescribeEngineDefaultParameters for more information about the operation and its arguments.

pub fn describe_events(&self) -> DescribeEvents<C, M, R>

Constructs a fluent builder for the DescribeEvents operation.

See DescribeEvents for more information about the operation and its arguments.

pub fn describe_global_replication_groups(
&self
) -> DescribeGlobalReplicationGroups<C, M, R>

Constructs a fluent builder for the DescribeGlobalReplicationGroups operation.

See DescribeGlobalReplicationGroups for more information about the operation and its arguments.

pub fn describe_replication_groups(&self) -> DescribeReplicationGroups<C, M, R>

Constructs a fluent builder for the DescribeReplicationGroups operation.

See DescribeReplicationGroups for more information about the operation and its arguments.

pub fn describe_reserved_cache_nodes(
&self
) -> DescribeReservedCacheNodes<C, M, R>

Constructs a fluent builder for the DescribeReservedCacheNodes operation.

See DescribeReservedCacheNodes for more information about the operation and its arguments.

pub fn describe_reserved_cache_nodes_offerings(
&self
) -> DescribeReservedCacheNodesOfferings<C, M, R>

Constructs a fluent builder for the DescribeReservedCacheNodesOfferings operation.

See DescribeReservedCacheNodesOfferings for more information about the operation and its arguments.

pub fn describe_service_updates(&self) -> DescribeServiceUpdates<C, M, R>

Constructs a fluent builder for the DescribeServiceUpdates operation.

See DescribeServiceUpdates for more information about the operation and its arguments.

pub fn describe_snapshots(&self) -> DescribeSnapshots<C, M, R>

Constructs a fluent builder for the DescribeSnapshots operation.

See DescribeSnapshots for more information about the operation and its arguments.

pub fn describe_update_actions(&self) -> DescribeUpdateActions<C, M, R>

Constructs a fluent builder for the DescribeUpdateActions operation.

See DescribeUpdateActions for more information about the operation and its arguments.

pub fn describe_user_groups(&self) -> DescribeUserGroups<C, M, R>

Constructs a fluent builder for the DescribeUserGroups operation.

See DescribeUserGroups for more information about the operation and its arguments.

pub fn describe_users(&self) -> DescribeUsers<C, M, R>

Constructs a fluent builder for the DescribeUsers operation.

See DescribeUsers for more information about the operation and its arguments.

pub fn disassociate_global_replication_group(
&self
) -> DisassociateGlobalReplicationGroup<C, M, R>

Constructs a fluent builder for the DisassociateGlobalReplicationGroup operation.

See DisassociateGlobalReplicationGroup for more information about the operation and its arguments.

pub fn failover_global_replication_group(
&self
) -> FailoverGlobalReplicationGroup<C, M, R>

Constructs a fluent builder for the FailoverGlobalReplicationGroup operation.

See FailoverGlobalReplicationGroup for more information about the operation and its arguments.

pub fn increase_node_groups_in_global_replication_group(
&self
) -> IncreaseNodeGroupsInGlobalReplicationGroup<C, M, R>

Constructs a fluent builder for the IncreaseNodeGroupsInGlobalReplicationGroup operation.

See IncreaseNodeGroupsInGlobalReplicationGroup for more information about the operation and its arguments.

pub fn increase_replica_count(&self) -> IncreaseReplicaCount<C, M, R>

Constructs a fluent builder for the IncreaseReplicaCount operation.

See IncreaseReplicaCount for more information about the operation and its arguments.

pub fn list_allowed_node_type_modifications(
&self
) -> ListAllowedNodeTypeModifications<C, M, R>

Constructs a fluent builder for the ListAllowedNodeTypeModifications operation.

See ListAllowedNodeTypeModifications for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn modify_cache_cluster(&self) -> ModifyCacheCluster<C, M, R>

Constructs a fluent builder for the ModifyCacheCluster operation.

See ModifyCacheCluster for more information about the operation and its arguments.

pub fn modify_cache_parameter_group(&self) -> ModifyCacheParameterGroup<C, M, R>

Constructs a fluent builder for the ModifyCacheParameterGroup operation.

See ModifyCacheParameterGroup for more information about the operation and its arguments.

pub fn modify_cache_subnet_group(&self) -> ModifyCacheSubnetGroup<C, M, R>

Constructs a fluent builder for the ModifyCacheSubnetGroup operation.

See ModifyCacheSubnetGroup for more information about the operation and its arguments.

pub fn modify_global_replication_group(
&self
) -> ModifyGlobalReplicationGroup<C, M, R>

Constructs a fluent builder for the ModifyGlobalReplicationGroup operation.

See ModifyGlobalReplicationGroup for more information about the operation and its arguments.

pub fn modify_replication_group(&self) -> ModifyReplicationGroup<C, M, R>

Constructs a fluent builder for the ModifyReplicationGroup operation.

See ModifyReplicationGroup for more information about the operation and its arguments.

pub fn modify_replication_group_shard_configuration(
&self
) -> ModifyReplicationGroupShardConfiguration<C, M, R>

Constructs a fluent builder for the ModifyReplicationGroupShardConfiguration operation.

See ModifyReplicationGroupShardConfiguration for more information about the operation and its arguments.

pub fn modify_user(&self) -> ModifyUser<C, M, R>

Constructs a fluent builder for the ModifyUser operation.

See ModifyUser for more information about the operation and its arguments.

pub fn modify_user_group(&self) -> ModifyUserGroup<C, M, R>

Constructs a fluent builder for the ModifyUserGroup operation.

See ModifyUserGroup for more information about the operation and its arguments.

pub fn purchase_reserved_cache_nodes_offering(
&self
) -> PurchaseReservedCacheNodesOffering<C, M, R>

Constructs a fluent builder for the PurchaseReservedCacheNodesOffering operation.

See PurchaseReservedCacheNodesOffering for more information about the operation and its arguments.

pub fn rebalance_slots_in_global_replication_group(
&self
) -> RebalanceSlotsInGlobalReplicationGroup<C, M, R>

Constructs a fluent builder for the RebalanceSlotsInGlobalReplicationGroup operation.

See RebalanceSlotsInGlobalReplicationGroup for more information about the operation and its arguments.

pub fn reboot_cache_cluster(&self) -> RebootCacheCluster<C, M, R>

Constructs a fluent builder for the RebootCacheCluster operation.

See RebootCacheCluster for more information about the operation and its arguments.

pub fn remove_tags_from_resource(&self) -> RemoveTagsFromResource<C, M, R>

Constructs a fluent builder for the RemoveTagsFromResource operation.

See RemoveTagsFromResource for more information about the operation and its arguments.

pub fn reset_cache_parameter_group(&self) -> ResetCacheParameterGroup<C, M, R>

Constructs a fluent builder for the ResetCacheParameterGroup operation.

See ResetCacheParameterGroup for more information about the operation and its arguments.

pub fn revoke_cache_security_group_ingress(
&self
) -> RevokeCacheSecurityGroupIngress<C, M, R>

Constructs a fluent builder for the RevokeCacheSecurityGroupIngress operation.

See RevokeCacheSecurityGroupIngress for more information about the operation and its arguments.

pub fn start_migration(&self) -> StartMigration<C, M, R>

Constructs a fluent builder for the StartMigration operation.

See StartMigration for more information about the operation and its arguments.

pub fn test_failover(&self) -> TestFailover<C, M, R>

Constructs a fluent builder for the TestFailover operation.

See TestFailover for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.